### 2025-06-24 1e786da
Update paper.md
