### 2025-06-24 #8473 @kyleniemeyer
@jorgepz It looks like one of your coauthors is missing an `affiliation` field in the paper and that is causing an error in the compilation; could you add this?
