### 2025-06-24 #8473 @jorgepz
Regarding the missing DOI error: [10.1016/0141-0296(84)90014-2](https://doi.org/10.1016/0141-0296(84)90014-2) is the doi of a review of the book we cited, not the book itself.
