### 2025-07-02 65fa789
agrego avance en unform curvature
