+++ examples/brinquedo/cantilever.m
@@ -105,7 +105,7 @@ Iy = Ly*tz^3/12 ;
 f_ana = qz*loadFactors_Shell*Ly*tz*Lx^3 / (3*E*Iy) 
 
 % =============================================================================================================================================
-stop
+% stop
 
 % geometrical scalar parameters
 l = Lx;
