+++ examples/uniformCurvatureCantilever/uniformCurvatureCantilever.m
@@ -35,7 +35,7 @@ end % hidden
 addpath(genpath([pwd '/../../src']));
 % material scalar parameters
 E = 200e9;
-nu = 0.3;
+nu = 0.0;
 % geometrical scalar parameters
 l = 10;
 ty = 1.0;
@@ -170,16 +170,6 @@ end
 [mesh.nodesCoords, mesh.conecCell] = meshFileReader([base_dir 'geometry_cantilever.msh']);
 assert(max(mesh.nodesCoords(:, 1)) == l && max(mesh.nodesCoords(:, 2)) == ty);
 
-% otherParams.problemName = 'uniformCurvatureCantilever-linearShell';
-
-% [modelCurrSol, modelProperties, BCsData] = initONSAS(materialsL, elements, boundaryConds, initialConds, mesh, analysisSettings, otherParams);
-
-% [matUs, loadFactorsMat, modelSolutions] = solveONSAS(modelCurrSol, modelProperties, BCsData);
-% %
-% controldofs = 6+[1 4 5] ;
-% controlDispShellLinear = -matUs(controldofs,:)
-% controlDispsNREngRot
-
 % ====================================================
 % shell non linear
 % ====================================================
@@ -188,60 +178,104 @@ otherParams.problemName = 'uniformCurvatureCantilever-nonLinearShell';
 %analysisSettings               = struct();
 %analysisSettings.methodName    = 'newtonRaphson';
 %analysisSettings.deltaT        =   0.;  % TEMPORARY
-%analysisSettings.finalTime      =   0.01;  % TEMPORARY
- %analysisSettings.finalTime      =   .0002;  % TEMPORARY
-%analysisSettings.stopTolDeltau =   1e-6;
-%analysisSettings.stopTolForces =   1e-4;
+analysisSettings.deltaT        =   0.01;  % TEMPORARY
+analysisSettings.finalTime      =   1;  % TEMPORARY
+% analysisSettings.finalTime      =   .0002;  % TEMPORARY
+% analysisSettings.stopTolDeltau =   1e-6;
+% analysisSettings.stopTolForces =   1e-6;
 analysisSettings.stopTolIts    =   15;
 [modelCurrSol, modelProperties, BCsData] = initONSAS(materialsNL, elements, boundaryConds, initialConds, mesh, analysisSettings, otherParams);
 [matUs, loadFactorsMat, modelSolutions] = solveONSAS(modelCurrSol, modelProperties, BCsData);
 
-controldofs = 3*6-2 ;
-controlDispShellNonLinear = -matUs(controldofs,:)
+% node = 3 ;
+node = 18 ;
+ux_dof_shell = node*6-5 ;
+uz_dof_shell = node*6-1 ;
+angle_dof_shell = node*6-2 ;
 
-controlDispShellNonLinear ./ controlDispShellLinear
-loadFactorsShell  =  loadFactorsMat(:, 2);
+% dofs_node2 = 2*6-2;
+% dofs_node3 = 3*6-2;
+% controlDofs_total = [dofs_node2; dofs_node3; controldofs] ;
+% angle_dof = -matUs(controlDofs_total,:)
+
+controlDispsShellNonLinear = -matUs(angle_dof_shell,:) ;
+control_ux_shell = matUs(ux_dof_shell,:) ;
+control_uz_shell = matUs(uz_dof_shell,:) ;
+
+% controlDispShellNonLinear ./ controlDispShellLinear
+loadFactorsShell  =  loadFactorsMat(:, 2)*ty*tz;
 %
+
[…]
