### 2025-07-02 0b3dc79
uniform curvature cantilever example
