+++ examples/uniformCurvatureCantilever/uniformCurvatureCantilever.m
@@ -34,12 +34,12 @@ end % hidden
 % add path
 addpath(genpath([pwd '/../../src']));
 % material scalar parameters
-E = 200e9;
-nu = 0.0;
+E = 200e9 ; % Young modulus
+nu = 0.0 ; % Poission 
 % geometrical scalar parameters
-l = 10;
-ty = 1.0;
-tz = .1;
+l = 10 ; % length
+ty = 1.0 ; % width
+tz = .1 ; % thickness
 Iy = ty * tz^3 / 12;
 Mobj = E * Iy * 2 * pi / l ;
 % the number of elements of the mesh
@@ -53,13 +53,6 @@ numElements = 10;
 % md## Numerical solution
 % md### MEB parameters
 % md
-
-materialsL                 = struct();
-materialsL.modelName  = 'elastic-linear';
-materialsL.modelParams = [E nu];
-
-
-
 % mdThe modelling of the structure begins with the definition of the Material-Element-BoundaryConditions (MEB) parameters.
 % md
 % md### materials
@@ -111,14 +104,13 @@ for i = 1:numElements
 end
 % md
 % md### analysisSettings
-analysisSettings               = struct();
+analysisSettings               = struct() ;
 analysisSettings.methodName    = 'newtonRaphson';
-analysisSettings.deltaT        =   0.1;  % TEMPORARY
-analysisSettings.finalTime      =   1;  % TEMPORARY
-% analysisSettings.finalTime      =   .0002;  % TEMPORARY
-analysisSettings.stopTolDeltau =   1e-6;
-analysisSettings.stopTolForces =   1e-6;
-analysisSettings.stopTolIts    =   10;
+analysisSettings.deltaT        =   0.1 ;
+analysisSettings.finalTime     =   1 ;  
+analysisSettings.stopTolDeltau =   1e-6 ;
+analysisSettings.stopTolForces =   1e-6 ;
+analysisSettings.stopTolIts    =   10 ;
 % md
 % md## otherParams
 otherParams             = struct();
@@ -127,33 +119,26 @@ otherParams.controlDofs = [numElements + 1  4];
 otherParams.plots_format = 'vtk';
 % md## Analysis case 1: NR with Rotated Eng Strain
 % md In the first case ONSAS is run and the solution at the dof (angle of node B) of interest is stored:
-
 [modelCurrSol, modelProperties, BCsData] = initONSAS(materialsNL, elements, boundaryConds, initialConds, mesh, analysisSettings, otherParams);
 %
 % mdAfter that the structs are used to perform the numerical time analysis
 [matUs, loadFactorsMat, modelSolutions] = solveONSAS(modelCurrSol, modelProperties, BCsData);
-
 % md
 % md the control dof to verificate the solution is the node angle B, this corresponds to the following dof number:
 angleControlDof      = (numElements + 1) * 6 - 2;
 controlDispsNREngRot =  -matUs(angleControlDof, :);
 loadFactorsNREngRot  =  loadFactorsMat(:, 2);
-% md and the analytical value of the load factors is computed
-analyticLoadFactorsNREngRot = @(w) E * Iy * w / l;
-% md
-
-
 
 % ====================================================
-% shell linear
+% shell non linear
 % ====================================================
-
+%
 elements             = struct();
[…]
