+++ examples/cantileverPlate/cantileverPlate.m
@@ -28,12 +28,21 @@ addpath(genpath([pwd '/../../src']));
 % md## Scalars
 E = 200e6;
 nu = 0.0;
-tz = .05;
-qx  = 100 ; % kN/m^2
-qz  = 10 ; % kN/m^2
-% md
 Ly = .5;
 Lx = 1;
+tz = .05;
+I = Ly * tz^3 / 12;
+%
+qx  = 100 ; % kN/m2
+qz  = 200 ; % kN/m2
+%
+Fz = qz * Ly * tz ;
+%
+% ====================================================
+% plate element
+% ====================================================
+%
+% md
 % md## Numerical solution using plate elements
 % md
 % md### Materials
@@ -58,11 +67,8 @@ boundaryConds(1).imposDispVals =  [0 0 0 0 0 0];
 %
 boundaryConds(2).loadsCoordSys = 'global';
 boundaryConds(2).loadsTimeFact = @(t) t;
-boundaryConds(2).loadsBaseVals = [0 0 0 0 -qz 0];
+boundaryConds(2).loadsBaseVals = [qx 0 0 0 -qz 0];
 %
-boundaryConds(3).loadsCoordSys = 'global';
-boundaryConds(3).loadsTimeFact = @(t) t;
-boundaryConds(3).loadsBaseVals = [qx 0 0 0 0 0];
 % md
 % md### mesh
 % md
@@ -91,128 +97,140 @@ analysisSettings.stopTolIts    =   10;
 otherParams                  = struct();
 % md The name of the problem is:
 % md
-% otherParams.problemName  = 'cantileverPlate-plateElem';
-% otherParams.plots_format = 'vtk';
+otherParams.problemName  = 'cantileverPlate-plateElem';
+otherParams.plots_format = 'vtk';
 % md
 % md Execute ONSAS and save the results:
-% [modelInitSol, modelProperties, BCsData] = initONSAS(materials, elements, boundaryConds, initialConds, mesh, analysisSettings, otherParams);
+[modelInitSol, modelProperties, BCsData] = initONSAS(materials, elements, boundaryConds, initialConds, mesh, analysisSettings, otherParams);
 %
 % mdAfter that the structs are used to perform the numerical time analysis
-% [matUs, loadFactorsMat, modelSolutions] = solveONSAS(modelInitSol, modelProperties, BCsData);
+[matUs, loadFactorsMat, modelSolutions] = solveONSAS(modelInitSol, modelProperties, BCsData);
 % md
 % md## verification
-% nelem = size(modelProperties.Conec, 1);
-% matSolic = getInternalForces(modelSolutions{end}.localInternalForces, 1:nelem, {'Mx', 'My', 'Mxy'});
-% numer_maxMx = max(max(matSolic));
-% numer_wmax = min(matUs(5:6:end));
-
+nelem = size(modelProperties.Conec, 1);
+matSolic = getInternalForces(modelSolutions{end}.localInternalForces, 1:nelem, {'Mx', 'My', 'Mxy'});
+numer_maxMx_plate = max(max(matSolic)) ;
+numer_wmax_plate = min(matUs(5:6:end)) ;
 % md
-analy_maxMx = qz * Lx / 2;
-qlin = qz * Ly;
-I = Ly * tz^3 / 12;
-analy_wmax = -qlin * Lx^4 / (8 * E * I);
-
-% elements(2).elemType           = 'triangle';
+analy_maxMx = Fz * Lx / Ly ;
+analy_wmax = -Fz * Lx^3 / (3 * E * I);
+analy_dxmax = qx * Lx / E;
+%
+% ====================================================
[…]
