+++ src/elements/triangle/internalForcesPlateTriangle.m
@@ -130,19 +130,17 @@ function [fs, ks, fintLocCoord, DD] = internalForcesPlateTriangle(elemCoords, el
 
   K = XYZ' * DD * XYZ;
 
-%   gamma_vec = [1 1 0];
-%   aux = zeros(3, 9);
-%   aux(1, 0 + (1:3)) = gamma_vec;
-%   aux(2, 3 + (1:3)) = gamma_vec;
-%   aux(3, 6 + (1:3)) = gamma_vec;
-%   vec_xis = 1 / DET * aux * (XYZ * elemDisps);
-%   Moments = 24 * D * vec_xis;
+  gamma_vec = [1 1 0];
+  aux = zeros(3, 9);
+  aux(1, 0 + (1:3)) = gamma_vec;
+  aux(2, 3 + (1:3)) = gamma_vec;
+  aux(3, 6 + (1:3)) = gamma_vec;
+  vec_xis = 1 / DET * aux * (XYZ * elemDisps);
+  Moments = 24 * D * vec_xis;
 
-%   fintLocCoord = Moments;
+  fintLocCoord = Moments;
 
-%   fint = K * elemDisps;
+  fint = K * elemDisps;
 
-%   fs = {fint};
-  fs = {0};
-    fintLocCoord = 0;
+  fs = {fint};
   ks = {K};
