### 2025-07-03 d2846ec
working on euler column example...
