### 2025-07-04 f19162c
fix deformation of shell and plate vtks
