+++ src/vtk/shellVtkData.m
@@ -29,6 +29,8 @@ function [vtkNodes, vtkConec, vtkNodalDisps, vtkInternalForces] ...
   % thickness
   tz = elemCrossSecParams{2};
 
+  Nodes = Nodes + reshape( U(1:2:end), 3, size(Nodes,1) )' ;
+
   indMx = 0;
   indMy = 0;
   indMxy = 0;
