+++ src/elements/triangle/co-rotational_funcs/matrixF.m
@@ -27,7 +27,7 @@ function [F1, F2] = matrixF(n, flag_second_mod) % ok
   n6 = n(16:18) ; 
   if flag_second_mod == 1
     % F1 = [ skew(n1)(1:2,:)' zeros(3,3) skew(n3)(1:2,:)' zeros(3,3) skew(n5)(1:2,:)' zeros(3,3) ]' ;
-    F1 = [ skew(n1)(1:2,:)' zeros(3,3) skew(n3)(1:2,:)' zeros(3,3) skew(n5)(1:2,:)' zeros(3,3) ]' ;
+    F1 = [ skew(n1)(1:2,:) ; zeros(3,3) ; skew(n3)(1:2,:) ; zeros(3,3) ; skew(n5)(1:2,:) ; zeros(3,3) ] ;
   else
     % F1 = [ skew(n1)' zeros(3,3) skew(n3)' zeros(3,3) skew(n5)' zeros(3,3) ]' ;
     F1 = [ skew(n1) ; zeros(3,3) ; skew(n3) ; zeros(3,3) ; skew(n5) ; zeros(3,3) ] ;
