+++ src/elements/triangle/co-rotational_funcs/d_Ts.m
@@ -17,49 +17,20 @@
 %
 function [d_Ts] = d_Ts(t, v)
 
-
-
-  % nt = norm(t);
-
-  % if nt == 0
-  %   d_Ts = 1 / 2 * skew(v);
-  % else
-  %   e = t / nt;
-  %   ev = cross(e, v);
-
-  %   a1 = (cos(nt) - sin(nt) / nt) / nt;
-  %   M1 = v * e' - (e' * v) * e * e';
-
-  %   a2 = (1 - sin(nt) / nt) / nt;
-  %   M2 = e * v' - 2 * (e' * v) * e * e' + (e' * v) * eye(3);
-
-  %   a3 = sin(nt) / nt - (2 * sin(nt / 2) / nt)^2;
-  %   M3 = ev * e';
-
-  %   a4 = 2 * (sin(nt / 2) / nt)^2;
-  %   M4 = skew(v);
-
-  %   d_Ts = a1 * M1 + a2 * M2 - a3 * M3 + a4 * M4;
-  % end
-
-% fprintf('=== \n')
-  % d_Ts
-  % fprintf('--------------------------------- \n')
   psi = norm(t);
   I = eye(3, 3);
 
   if psi == 0
-    d_Ts = 1 / 2 * skew(v);
+    d_Ts = 0.5 * skew(v);
   else
-    u = t / psi ;
-    a = sin(psi) / psi ;
-    b = ( sin(psi/2) / (psi/2) ) ;
-    sum1 = -(a - b^2) * cross(u,v) * u' ;
-    sum2 = 1/2*b^2 * skew(v) ;
-    sum3 = ( cos(psi) -a ) * 1/psi * (v*u'- (u'*v)*u*u' ) ;
-    sum4 = (1-a)* 1/psi * ( u*v' -2*(u'*v)*u*u' + (u'*v)*I ) ; 
-    d_Ts = sum1 + sum2 + sum3 + sum4 ;
+    u = t / psi;
+    a = sin(psi) / psi;
+    b = (sin(psi/2) / (psi/2));
+    sum1 = -(a - b^2) * cross(u,v) * u';
+    sum2 = 0.5 * b^2 * skew(v);
+    sum3 = (cos(psi)-a) * 1 / psi * (v * u'- (u' * v) * u * u');
+    sum4 = (1-a) * 1 / psi * (u * v' -2 * (u' * v) * u * u' + (u' * v) * I); 
+    d_Ts = sum1 + sum2 + sum3 + sum4;
   end
-  % d_Ts
-  % fprintf('=== \n')
+
 end
