+++ src/elements/triangle/internalForcesShellTriangle.m
@@ -32,7 +32,7 @@ function [fs, ks, fintLocCoord, rotMat] = internalForcesShellTriangle(elemCoords
   poisson_ratio = modelParams(2);
   h = thickness;
 
-  % Nodes position vector in global reference frame 
+  % Nodes position vector in global reference frame
   elemCoords = elemCoords';
   r1_g = elemCoords(1:3);
   r2_g = elemCoords(4:6);
@@ -49,8 +49,8 @@ function [fs, ks, fintLocCoord, rotMat] = internalForcesShellTriangle(elemCoords
   u2_g    = Ug(7:9);
   u3_g    = Ug(13:15);
   t1_g    = Ug(4:6);
-  t2_g    = Ug(10:12) ;
-  t3_g    = Ug(16:18) ;
+  t2_g    = Ug(10:12);
+  t3_g    = Ug(16:18);
 
   % Updated position vector in global reference frame
   p1_g = r1_g + u1_g;
@@ -62,7 +62,7 @@ function [fs, ks, fintLocCoord, rotMat] = internalForcesShellTriangle(elemCoords
     pog = p1_g;
   end
   % fprintf('p_g \n')
-  % [ p1_g p2_g p3_g pog ] 
+  % [ p1_g p2_g p3_g pog ]
   
   % Global rotation matrix
   % eq. (35) of 10.1016/j.cma.2006.10.006
@@ -115,16 +115,16 @@ function [fs, ks, fintLocCoord, rotMat] = internalForcesShellTriangle(elemCoords
     a   = [a1_def, a2_def, a3_def];
     ro  = [r1_o, r2_o, r3_o];
     for i = 1:3
-      ai = a(:,i);
-      rio = ro(:,i);
+      ai = a(:, i);
+      rio = ro(:, i);
       auxNum = ai(2) * rio(1) - ai(1) * rio(2);
       Num = Num + auxNum;
 
-      auxDen = ai(1) * rio(1) + ai(2) * rio(2) ;
+      auxDen = ai(1) * rio(1) + ai(2) * rio(2);
       Den = Den + auxDen;
     end  
-    tan_theta = Num/Den 
-    theta = rad2deg(atan(tan_theta))
+    tan_theta = Num / Den;
+    theta = rad2deg(atan(tan_theta));
     % num = a1_def(2)*r1_o(1) - a1_def(1)*r1_o(2) + a2_def(2)*r2_o(1) - a2_def(1)*r2_o(2) + a3_def(2)*r3_o(1) - a3_def(1)*r3_o(2)
     % den = a1_def(1)*r1_o(1) + a1_def(2)*r1_o(2) + a2_def(1)*r2_o(1) + a2_def(2)*r2_o(2) + a3_def(1)*r3_o(1) + a3_def(2)*r3_o(2)
     % stop
@@ -137,14 +137,14 @@ function [fs, ks, fintLocCoord, rotMat] = internalForcesShellTriangle(elemCoords
   % [ sum(G(:,1)) sum(G(:,2)) sum(G(:,3)) ]
   % stop
 
-  % Rotation matrix from local reference frame to nodal reference frame in deformed configuration 
+  % Rotation matrix from local reference frame to nodal reference frame in deformed configuration
   % eq. (2) of 10.1016/j.cma.2006.10.006
   R1_def = Rr' * R1_g * Ro;
   R2_def = Rr' * R2_g * Ro;
   R3_def = Rr' * R3_g * Ro;
   % stop
 
-  % Nodal rotations in local reference frame in deformed configuration 
[…]
