+++ src/elements/triangle/internalForcesLinearShellTriangle.m
@@ -19,7 +19,7 @@
 % The element is formed by the superposition of a plate element (DKT) and a plane stress element (CST) with
 % with addition to artificial drilling (rotation about the axis normal to the element plane) stiffness.
 %
-function [fs, ks, fintLocCoord,Kefora, Kb ] = internalForcesLinearShellTriangle(elemCoords, elemDisps, modelName, modelParams, thickness)
+function [fs, ks, fintLocCoord, Kefora, Kb] = internalForcesLinearShellTriangle(elemCoords, elemDisps, modelName, modelParams, thickness)
 
   % material and geometric parameters
   young_modulus = modelParams(1);
@@ -37,14 +37,11 @@ function [fs, ks, fintLocCoord,Kefora, Kb ] = internalForcesLinearShellTriangle(
 
   [T, x02, x03, y03] = edgeLocalAxisShellTriangle(r1g, r2g, r3g);
   Te = blkdiag(T, T, T, T, T, T);
-
   Ul = Te * Ug;
 
   [Kl, fintLocCoord, Kb] = localShellTriangle(x02, x03, y03, young_modulus, poisson_ratio, h, Ul);
 
-
   Kg = Te' * Kl * Te;
-
   fg = Kg * Ug;
 
   ks = {switchToNodalIndexing(Kg)};
