+++ src/elements/triangle/co-rotational_funcs/matrixGi.m
@@ -23,48 +23,48 @@ function [G1, G2, G3] = matrixGi(a1, a2, a3, r1, r2, r3, e1_flag, origin_flag) %
   a13 = a1 - a3;
 
   % if origin_flag == 0
-    % Eq. (27) of 10.1016/j.cma.2006.10.006
-    
-    d = norm(a21);
-    v = norm(cross(a21, a31));
-    c = a1' * r1 + a2' * r2 + a3' * r3;
+  % Eq. (27) of 10.1016/j.cma.2006.10.006
+  
+  d = norm(a21);
+  v = norm(cross(a21, a31));
+  c = a1' * r1 + a2' * r2 + a3' * r3;
 
-    G1 = zeros(6, 3);
-    G1(3, 1) = a32(1) / v;
-    G1(3, 2) = a32(2) / v;
+  G1 = zeros(6, 3);
+  G1(3, 1) = a32(1) / v;
+  G1(3, 2) = a32(2) / v;
 
-    G2 = zeros(6, 3);
-    G2(3, 1) = a13(1) / v;
-    G2(3, 2) = a13(2) / v;
-    
-    G3 = zeros(6, 3);
-    G3(3, 1) = a21(1) / v;
-    G3(3, 2) = a21(2) / v;
-    % G3(3, 1) = a32(1) / v;
-    % G3(3, 2) = a32(2) / v;
+  G2 = zeros(6, 3);
+  G2(3, 1) = a13(1) / v;
+  G2(3, 2) = a13(2) / v;
+  
+  G3 = zeros(6, 3);
+  G3(3, 1) = a21(1) / v;
+  G3(3, 2) = a21(2) / v;
+  % G3(3, 1) = a32(1) / v;
+  % G3(3, 2) = a32(2) / v;
 
-    % e1_flag
-    if e1_flag == 0
-      %
-      G1(1, 3) =  -r1(2) / c;
-      G1(2, 3) =   r1(1) / c;
-      %
-      G2(1, 3) =  -r2(2) / c;
-      G2(2, 3) =   r2(1) / c;
-      %
-      G3(1, 3) =  -r3(2) / c;
-      G3(2, 3) =   r3(1) / c;
-    
-    else
-      G1(1, 3) = 0;
-      G1(2, 3) = -1/d;
-      %
-      G2(1, 3) = 0;
-      G2(2, 3) = 1/d;
-      %
-      % G3(1, 3) = 0;
-      % G3(2, 3) = 0;    
-    end
+  % e1_flag
+  if e1_flag == 0
+    %
+    G1(1, 3) =  -r1(2) / c;
+    G1(2, 3) =   r1(1) / c;
+    %
+    G2(1, 3) =  -r2(2) / c;
+    G2(2, 3) =   r2(1) / c;
+    %
+    G3(1, 3) =  -r3(2) / c;
+    G3(2, 3) =   r3(1) / c;
+  
+  else
+    G1(1, 3) = 0;
+    G1(2, 3) = -1/d;
+    %
+    G2(1, 3) = 0;
+    G2(2, 3) = 1/d;
+    %
+    % G3(1, 3) = 0;
+    % G3(2, 3) = 0;    
+  end
 
   % elseif origin_flag == 1
 
