+++ src/elements/triangle/co-rotational_funcs/matrixKhi.m
@@ -23,4 +23,3 @@ function [Khi] = matrixKhi(R, m) % ok
                [(R(3, 3) * m(2) - R(3, 2) * m(3)),                          (R(3, 1) * m(3) - R(3, 3) * m(1)),                            ((R(1, 2) - R(2, 1)) * m(3) + R(2, 3) * m(1) - R(1, 3) * m(2))]];
 
 end
-
