+++ src/interface/initONSAS.m
@@ -103,9 +103,9 @@ function [modelCurrSol, modelProperties, BCsData] = initONSAS(materials, element
   % ==========================================================================================
   % ==========================================================================================
   % Global rotation matrices for triangle elements
-  rotMatCell = cell(size(U, 1)/6);
+  rotMatCell = cell(size(U, 1) / 6);
   % Esto se tiene que mejorar, se asume que no hay rotaciones iniciales
-  rotMatCell(:) = {eye(3)} ; 
+  rotMatCell(:) = {eye(3)}; 
   % ==========================================================================================
   % ==========================================================================================
 
