+++ src/elements/triangle/co-rotational_funcs/matrixTm.m
@@ -22,8 +22,8 @@ function [Tmi] = matrixTm(qi) % ok
   q3 = qi(3);
   q0 = sqrt(1.0 - q1^2 - q2^2 - q3^2);
 
-  Tmi = 2 / q0 * [[(q0^2 + q1^2)      ,  (q1 * q2 - q0 * q3)  ,  (q1 * q3 + q0 * q2)]
-                 [(q1 * q2 + q0 * q3) ,  (q0^2 + q2^2)        ,  (q2 * q3 - q0 * q1)]
-                 [(q1 * q3 - q0 * q2) ,  (q2 * q3 + q0 * q1)  ,  (q0^2 + q3^2)      ]
+  Tmi = 2 / q0 * [[(q0^2 + q1^2),       (q1 * q2 - q0 * q3),  (q1 * q3 + q0 * q2)]
+                 [(q1 * q2 + q0 * q3),  (q0^2 + q2^2),        (q2 * q3 - q0 * q1)]
+                 [(q1 * q3 - q0 * q2),  (q2 * q3 + q0 * q1),  (q0^2 + q3^2)      ]
                  ];
 end

[…]
