+++ src/elements/triangle/co-rotational_funcs/matrixTm.m
@@ -23,7 +23,7 @@ function [Tmi] = matrixTm(qi) % ok
   q0 = sqrt(1.0 - q1^2 - q2^2 - q3^2);
 
   Tmi = 2 / q0 * [[(q0^2 + q1^2),       (q1 * q2 - q0 * q3),  (q1 * q3 + q0 * q2)]
-                 [(q1 * q2 + q0 * q3),  (q0^2 + q2^2),        (q2 * q3 - q0 * q1)]
-                 [(q1 * q3 - q0 * q2),  (q2 * q3 + q0 * q1),  (q0^2 + q3^2)      ]
+                  [(q1 * q2 + q0 * q3),  (q0^2 + q2^2),        (q2 * q3 - q0 * q1)]
+                  [(q1 * q3 - q0 * q2),  (q2 * q3 + q0 * q1),  (q0^2 + q3^2)]
                  ];
 end
