+++ examples/uniformCurvatureCantilever/uniformCurvatureCantilever.m
@@ -202,7 +202,7 @@ verifBoolean = norm(analyticLoadFactorsNREngRot(controlDispsNREngRot) - ...
                     loadFactorsNREngRot')  < ...
               (norm(analyticLoadFactorsNREngRot(controlDispsNREngRot)) * 1e-4) && ...
               norm(analyticLoadFactorsNREngRot(controlDispsShellNonLinear) - ...
-                  loadFactorsShell')  < ...
+                   loadFactorsShell')  < ...
               (norm(analyticLoadFactorsNREngRot(controlDispsShellNonLinear)) * 1e-2);
 %
 close all;
