### 2025-07-07 26a2a69
fix rotations shell
