+++ src/vtk/shellVtkData.m
@@ -29,6 +29,7 @@ function [vtkNodes, vtkConec, vtkNodalDisps, vtkInternalForces] ...
   % thickness
   tz = elemCrossSecParams{2};
 
+  Nodes_ref = Nodes;
   Nodes = Nodes + reshape(U(1:2:end), 3, size(Nodes, 1))';
 
   indMx = 0;
@@ -53,33 +54,32 @@ function [vtkNodes, vtkConec, vtkNodalDisps, vtkInternalForces] ...
     nodesElem  = Conec(i, 1:3);
     dofsElem   = nodes2dofs(nodesElem, 6);
 
+    % column vector with displacements of the dofs of the current element
+    dispsElem  = U(dofsElem);
+    aux  = reshape(dispsElem(1:2:end)', [3, 3])';
+    rots = reshape(dispsElem(2:2:end)', [3, 3])';
+
     crossVector = cross( ...
-                        Nodes(nodesElem(2), :) - Nodes(nodesElem(1), :), ...
-                        Nodes(nodesElem(3), :) - Nodes(nodesElem(1), :) ...
+                        Nodes_ref(nodesElem(2), :) - Nodes_ref(nodesElem(1), :), ...
+                        Nodes_ref(nodesElem(3), :) - Nodes_ref(nodesElem(1), :) ...
                        );
     normalVector = crossVector / norm(crossVector);
 
     coordsThickness = tz * .5 * [ones(3, 1); -ones(3, 1)] * normalVector;
 
-    Nodesvtk = [Nodes(nodesElem, :);   ...
-                Nodes(nodesElem, :)] + ...
-               coordsThickness;
-
-    % column vector with displacements of the dofs of the current element
-    dispsElem  = U(dofsElem);
-
-    aux  = reshape(dispsElem(1:2:end)', [3, 3])';
-    rots = reshape(dispsElem(2:2:end)', [3, 3])';
-
-    Dispsvtk = ...
-     [(expon(rots(1, :)) * coordsThickness(1, :)')'; ...
+    defThickness = ...
+    [(expon(rots(1, :)) * coordsThickness(1, :)')'; ...
       (expon(rots(2, :)) * coordsThickness(2, :)')'; ...
       (expon(rots(3, :)) * coordsThickness(3, :)')'; ...
       (expon(rots(1, :)) * coordsThickness(4, :)')'; ...
       (expon(rots(2, :)) * coordsThickness(5, :)')'; ...
-      (expon(rots(3, :)) * coordsThickness(6, :)')'] - ...
-     coordsThickness + ...
-     [aux; aux];
+      (expon(rots(3, :)) * coordsThickness(6, :)')'] ;
+
+    Nodesvtk = [Nodes(nodesElem, :);   ...
+                Nodes(nodesElem, :)] + ...
+               defThickness;
+
+    Dispsvtk = defThickness - coordsThickness + [aux; aux];
 
     Conecvtk = [13  (nodes2dofs(i, 6) - 1)'];
 
