+++ src/vtk/shellVtkData.m
@@ -68,12 +68,12 @@ function [vtkNodes, vtkConec, vtkNodalDisps, vtkInternalForces] ...
     coordsThickness = tz * .5 * [ones(3, 1); -ones(3, 1)] * normalVector;
 
     defThickness = ...
-    [(expon(rots(1, :)) * coordsThickness(1, :)')'; ...
+     [(expon(rots(1, :)) * coordsThickness(1, :)')'; ...
       (expon(rots(2, :)) * coordsThickness(2, :)')'; ...
       (expon(rots(3, :)) * coordsThickness(3, :)')'; ...
       (expon(rots(1, :)) * coordsThickness(4, :)')'; ...
       (expon(rots(2, :)) * coordsThickness(5, :)')'; ...
-      (expon(rots(3, :)) * coordsThickness(6, :)')'] ;
+      (expon(rots(3, :)) * coordsThickness(6, :)')'];
 
     Nodesvtk = [Nodes(nodesElem, :);   ...
                 Nodes(nodesElem, :)] + ...
