+++ src/core/convergenceTest.m
@@ -17,7 +17,6 @@
 %
 function [booleanConverged, stopCritPar, deltaErrLoad, normFext] = convergenceTest( ...
                                                                                    analysisSettings, redFext, redDeltaU, redUk, dispIter, systemDeltauRHS)
-  % dispIter
   stopTolDeltau = analysisSettings.stopTolDeltau;
   stopTolForces = analysisSettings.stopTolForces;
   stopTolIts    = analysisSettings.stopTolIts;
@@ -29,7 +28,6 @@ function [booleanConverged, stopCritPar, deltaErrLoad, normFext] = convergenceTe
   normFext      = norm(redFext);
 
   logicDispStop = (normadeltau  < (normaUk  * stopTolDeltau));
-
   logicForcStop = (deltaErrLoad < ((normFext + (normFext < stopTolForces)) * stopTolForces))  * (deltaErrLoad > 0);
 
   if isnan(norm(redDeltaU))
