+++ test/runTestProblemsLocal.m
@@ -1,23 +1,3 @@
-% Copyright 2025, ONSAS Authors (see documentation)
-%
-% This file is part of ONSAS.
-%
-% ONSAS is free software: you can redistribute it and/or modify
-% it under the terms of the GNU General Public License as published by
-% the Free Software Foundation, either version 3 of the License, or
-% (at your option) any later version.
-%
-% ONSAS is distributed in the hope that it will be useful,
-% but WITHOUT ANY WARRANTY; without even the implied warranty of
-% MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-% GNU General Public License for more details.
-%
-% You should have received a copy of the GNU General Public License
-% along with ONSAS.  If not, see <https://www.gnu.org/licenses/>.
-%
-% This functions reads a text file with the norms of the forces at
-% each iteration of the code execution and generates a series of plots
-
 close all;
 clear all;
 addpath(genpath([pwd filesep '..' filesep  'src' filesep]));
