+++ docs/src/bringONSASmFilesToONSASdocs.m
@@ -1,25 +1,3 @@
-% Copyright 2025, ONSAS Authors (see documentation)
-%
-% This file is part of ONSAS.
-%
-% ONSAS is free software: you can redistribute it and/or modify
-% it under the terms of the GNU General Public License as published by
-% the Free Software Foundation, either version 3 of the License, or
-% (at your option) any later version.
-%
-% ONSAS is distributed in the hope that it will be useful,
-% but WITHOUT ANY WARRANTY; without even the implied warranty of
-% MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-% GNU General Public License for more details.
-%
-% You should have received a copy of the GNU General Public License
-% along with ONSAS.  If not, see <https://www.gnu.org/licenses/>.
-%
-% md function used to convert .m files in the ONSAS repo to .md files in this repo
-% md this function is executed from the makeAndPreview.sh file or can also be
-% md executed from the bash terminal: `octave --eval "bringONSASmFilesToONSAS_docs('$ONSAS_PATH')"`
-% md where \$ONSAS_PATH is the environment variable with the directory of ONSAS
-
 function bringONSASmFilesToONSASdocs()
   disp('running bringONSASmFilesToONSASdocs script...');
   dirONSASdocs = './examples/';
