### 2025-07-30 #8473 @jorgepz
Hi @kyleniemeyer  just wanted to let you know that, the development of the package continues and, we have just done a minor release (now v0.4.0) and also fixed an issue about github detecting "M" code while it is only Matlab.

I was wondering if it is possible to remove the "M" code label, and also I am not sure if it would be needed to update the version in the description of the current issue (which refers to version 0.3.3 of the master 5 weeks ago).
