+++ src/core/computeDeltaU.m
@@ -51,12 +51,12 @@ if strcmp( analysisSettings.methodName, 'arcLength' )
 
       % Follow the sign of the predictor work increment (incremental work)
       % Fext = systemDeltauMatrix * (arcLengthNorm .* deltaubar) ;
-      % sign(deltaubar'*Fext)
+      % sign(deltaubar'*Fext) ;
       % deltalambda(1) = sign(deltaubar'*Fext) * incremArcLen / sqrt( deltaubar' * ( arcLengthNorm .* deltaubar ) ) ;
      
       % Follow the sign of the stiffness determinant
       % detKT = det(systemDeltauMatrix) ;
-      % sign(detKT)
+      % sign(detKT) ;
       % deltalambda(1) = sign(detKT) * incremArcLen / sqrt( deltaubar' * ( arcLengthNorm .* deltaubar ) ) ;
     
     end
