+++ examples/cantileverPlate/pruebas.m
@@ -1,0 +1,52 @@
+clc ; clear all ; close all
+
+addpath(genpath([pwd '/../../src']));
+
+
+E = 120;
+nu = 0.25;
+
+h = 1/8;
+
+r1 = [0,0,0];
+r2 = [4.08, -3.44,0];
+r3 = [3.4, 1.14, 0];
+
+U = zeros(18,1) ;
+
+[~, x02, ~, y03] = edgeLocalAxisShellTriangle(r1, r2, r3);
+A = x02 * y03 / 2;
+
+[Km, Kb, Kh] = opt_membrane_element(U, E, nu, h, A, r1, r2, r3);
+
+
+Kb_paper = [
+    10.350 0.95258 7.7327 -2.1309 1.7629 2.1745 -8.2194 -2.7155 -9.9073
+ 0.95258 4.0758 8.1695 2.7629 0.17327 -5.7458 -3.7155 -4.2491 -2.4237
+ 7.7327 8.1695 19.723 3.9414 1.3377 -9.4943 -11.674 -9.5072 -10.229
+ -2.1309 2.7629 3.9414 2.7575 -1.1855 -4.1179 -0.62660 -1.5774 0.17657
+ 1.7629 0.17327 1.3377 -1.1855 5.8957 -4.6390 -0.57737 -6.0690 3.3013
+ 2.1745 -5.7458 -9.4943 -4.1179 -4.6390 13.777 1.9434 10.385 -4.2825
+ -8.2194 -3.7155 -11.674 -0.62660 -0.57737 1.9434 8.8460 4.2928 9.7307
+ -2.7155 -4.2491 -9.5072 -1.5774 -6.0690 10.385 4.2928 10.318 -0.87754
+ -9.9073 -2.4237 -10.229 0.17657 3.3013 -4.2825 9.7307 -0.87754 14.512
+];
+
+Kb./Kb_paper;
+
+Kh
+Kh_paper = [
+    0.041538 -0.27977 -0.63728 0.20769 0.069638 -0.52781 -0.24923 0.21014 -0.83208
+ -0.27977 1.8844 4.2923 -1.3989 -0.46903 3.5549 1.6786 -1.4153 5.6043
+ -0.63728 4.2923 12.833 -3.1864 -1.0684 7.7151 3.8237 -3.2239 10.092
+ 0.20769 -1.3989 -3.1864 1.0385 0.34819 -2.6390 -1.2462 1.0507 -4.1604
+ 0.069638 -0.46903 -1.0684 0.34819 0.11675 -0.88485 -0.41783 0.35229 -1.3950
+ -0.52781 3.5549 7.7151 -2.6390 -0.88485 7.9515 3.1668 -2.6701 9.7104
+ -0.24923 1.6786 3.8237 -1.2462 -0.41783 3.1668 1.4954 -1.2608 4.9925
+ 0.21014 -1.4153 -3.2239 1.0507 0.35229 -2.6701 -1.2608 1.0630 -4.2093
+ -0.83208 5.6043 10.092 -4.1604 -1.3950 9.7104 4.9925 -4.2093 20.204
+];
+
+Kh./Kh_paper
+
+% Km
