+++ src/elements/triangle/localShellTriangle.m
@@ -15,7 +15,7 @@
 % You should have received a copy of the GNU General Public License
 % along with ONSAS.  If not, see <https://www.gnu.org/licenses/>.
 %
-function [Kel, fintLocCoord, Kb] = localShellTriangle(x02, x03, y03, E, nu, h, Ul, flag_OPT)
+function [Kel, fintLocCoord, Kb] = localShellTriangle(x02, x03, y03, E, nu, h, Ul, flag_OPT, r1_g, r2_g, r3_g, Tr)
 
   % Calculate the area of the triangle
   area = x02 * y03 / 2;
@@ -32,7 +32,12 @@ function [Kel, fintLocCoord, Kb] = localShellTriangle(x02, x03, y03, E, nu, h, U
 	Ulm = Ul(im);
 	N = Dm * Bm * Ulm;
   else
-    Km = opt_membrane_element() ;
+	Km_opt = opt_membrane_element(E, nu, h, area, r1_g, r2_g, r3_g) ;
+	im = [1, 2, 6, 7, 8, 12, 13, 14, 18];
+	%
+	% Te = blkdiag(Tr, Tr, Tr);
+	% Km = Te * Km_opt * Te' ;
+	Km = Km_opt;
   end
 
   % bending stiffness
@@ -60,14 +65,14 @@ function [Kel, fintLocCoord, Kb] = localShellTriangle(x02, x03, y03, E, nu, h, U
   Kel(im, im) = Km;
   Kel(ib, ib) = Kb;
 
-  k_dr = min(min(abs(Kb))) * 1.e-4;
-  % k_dr = max(k_dr, 1.e-4);
-  % k_dr=0;
-  Kel(6, 6) = k_dr;
-  Kel(12, 12) = k_dr;
-  Kel(18, 18) = k_dr;
-
-  Kel = real(Kel);
+  if flag_OPT == 0
+	k_dr = min(min(abs(Kb))) * 1.e-4;
+	% k_dr = max(k_dr, 1.e-4);
+	% k_dr=0;
+	Kel(6, 6) = k_dr;
+	Kel(12, 12) = k_dr;
+	Kel(18, 18) = k_dr;
+  end
 
   % returning moments in local element coordiantes
   fintLocCoord = zeros(1, 3);
