+++ src/elements/triangle/localShellTriangle.m
@@ -15,31 +15,13 @@
 % You should have received a copy of the GNU General Public License
 % along with ONSAS.  If not, see <https://www.gnu.org/licenses/>.
 %
-function [Kel, fintLocCoord, Kb] = localShellTriangle(x02, x03, y03, E, nu, h, Ul, flag_OPT, r1_g, r2_g, r3_g, Tr)
+function [Kel, fintLocCoord] = localShellTriangle(x02, x03, y03, E, nu, h, Ul, flag_OPT, r1_g, r2_g, r3_g, To)
 
   % Calculate the area of the triangle
   area = x02 * y03 / 2;
-
-  % membrane stiffness
-  if flag_OPT == 0
-	aux1 = h *  E / (1 - nu^2);
-	aux2 = nu * aux1;
-	Dm = [[aux1, aux2, 0]; [aux2, aux1, 0]; [0, 0, aux1 * (1 - nu) / 2]];
-	Bm = cstB(x02, x03, y03);
-	Km = area * Bm' * Dm * Bm;
-	im = [1, 2, 7, 8, 13, 14];
-	% membrane forces (constant)
-	Ulm = Ul(im);
-	N = Dm * Bm * Ulm;
-  else
-	Km_opt = opt_membrane_element(E, nu, h, area, r1_g, r2_g, r3_g) ;
-	im = [1, 2, 6, 7, 8, 12, 13, 14, 18];
-	%
-	% Te = blkdiag(Tr, Tr, Tr);
-	% Km = Te * Km_opt * Te' ;
-	Km = Km_opt;
-  end
-
+  %
+  Kel = zeros(18, 18);
+  %
   % bending stiffness
   aux1 = E * h^3 / (12 * (1 - nu^2));
   aux2 = nu * aux1;
@@ -60,19 +42,31 @@ function [Kel, fintLocCoord, Kb] = localShellTriangle(x02, x03, y03, E, nu, h, U
   curv = Bb * Ulb;
   M = Db * curv;
 
+  % membrane stiffness
+  if flag_OPT == 0
+    aux1 = h *  E / (1 - nu^2);
+    aux2 = nu * aux1;
+    Dm = [[aux1, aux2, 0]; [aux2, aux1, 0]; [0, 0, aux1 * (1 - nu) / 2]];
+    Bm = cstB(x02, x03, y03);
+    Km = area * Bm' * Dm * Bm;
+    im = [1, 2, 7, 8, 13, 14];
+    % membrane forces (constant)
+    Ulm = Ul(im);
+    N = Dm * Bm * Ulm;
+    k_dr = min(min(abs(Kb))) * 1.e-4;
+    Kel(6, 6) = k_dr;
+    Kel(12, 12) = k_dr;
+    Kel(18, 18) = k_dr;
+  else
+    Km_opt = opt_membrane_element(E, nu, h, area, r1_g, r2_g, r3_g) ;
+    im = [1, 2, 6, 7, 8, 12, 13, 14, 18];
+    Te = blkdiag(To, To, To);
+    Km = Te * Km_opt * Te' ;    
+  end
+
   % assembling the stiffness matrix of the shell element in local coordinates
-  Kel = zeros(18, 18);
   Kel(im, im) = Km;
   Kel(ib, ib) = Kb;
-
-  if flag_OPT == 0
-	k_dr = min(min(abs(Kb))) * 1.e-4;
-	% k_dr = max(k_dr, 1.e-4);
-	% k_dr=0;
-	Kel(6, 6) = k_dr;
-	Kel(12, 12) = k_dr;
-	Kel(18, 18) = k_dr;
-  end
 
   % returning moments in local element coordiantes
   fintLocCoord = zeros(1, 3);
