+++ src/elements/triangle/internalForcesShellTriangle.m
@@ -26,7 +26,7 @@ function [fs, ks, fintLocCoord, rotMat] = internalForcesShellTriangle(elemCoords
   flag_first_mod  = 0; % 1 if battini modification is used - 0 if not / local rotations
   flag_second_mod = 0;  % 1 if battini modification is used - 0 if not / out of plane disps = 0
   flag_third_mod  = 0;  % 1 if battini modification is used - 0 if not / quaternions
-  flag_OPT        = 1;  % 1 if battini modification is used - 0 if not / quaternions
+  flag_OPT        = 0;  % 1 if battini modification is used - 0 if not / quaternions
 
   % material and geometric parameters
   young_modulus = modelParams(1);
@@ -62,21 +62,12 @@ function [fs, ks, fintLocCoord, rotMat] = internalForcesShellTriangle(elemCoords
   elseif origin == 1
     pog = p1_g;
   end
-  % fprintf('p_g \n')
-  % [ p1_g p2_g p3_g pog ]
 
   % Global rotation matrix
   % eq. (35) of 10.1016/j.cma.2006.10.006
-  % R1_g = rotMat{1};
-  % R2_g = rotMat{2};
-  % R3_g = rotMat{3};
   R1_g = expm(skew(t1_g));
   R2_g = expm(skew(t2_g));
   R3_g = expm(skew(t3_g));
-  % R1_g = expon(t1_g);
-  % R2_g = expon(t2_g);
-  % R3_g = expon(t3_g);
-  % stop
 
   % Transformation matrices from global reference frame
   [To, x02, x03, y03] = edgeLocalAxisShellTriangle(r1_g, r2_g, r3_g);
@@ -92,21 +83,15 @@ function [fs, ks, fintLocCoord, rotMat] = internalForcesShellTriangle(elemCoords
   r1_o = To * (r1_g - rc_g);
   r2_o = To * (r2_g - rc_g);
   r3_o = To * (r3_g - rc_g);
-  % fprintf('ri_o \n')
-  % [ r1_o r2_o r3_o ]
 
   u1_def = Rr' * (p1_g - pog) - r1_o;
   u2_def = Rr' * (p2_g - pog) - r2_o;
   u3_def = Rr' * (p3_g - pog) - r3_o;
-  % fprintf('u_def \n')
-  % [ u1_def u2_def u3_def ]
-  % stop
 
   % eq. (7) of 10.1016/j.cma.2006.10.006
   a1_def = u1_def + r1_o;
   a2_def = u2_def + r2_o;
   a3_def = u3_def + r3_o;
-  % [a1_def a2_def a3_def]
 
   if e1_parallel_side12 == 0
     Num = 0;
@@ -133,33 +118,26 @@ function [fs, ks, fintLocCoord, rotMat] = internalForcesShellTriangle(elemCoords
   % eq. (27) of 10.1016/j.cma.2006.10.006
   [G1, G2, G3] = matrixGi(a1_def, a2_def, a3_def, r1_o, r2_o, r3_o, e1_parallel_side12, origin);
   G = [G1; G2; G3];
-  % [ sum(G(:,1)) sum(G(:,2)) sum(G(:,3)) ]
-  % stop
 
   % Rotation matrix from local reference frame to nodal reference frame in deformed configuration
   % eq. (2) of 10.1016/j.cma.2006.10.006
   R1_def = Rr' * R1_g * Ro;
   R2_def = Rr' * R2_g * Ro;
   R3_def = Rr' * R3_g * Ro;
-  % stop
 
[…]
