### 2025-09-08 #8473 @jorgepz
Dear @kyleniemeyer, I hope you are well.

I'm writing to ask for your guidance on the next steps for this submission. The paper seems to be a great fit for the journal, and I am very keen for JOSS to publish it. However, we have not yet been able to secure an editor.

As I'm not familiar with the typical timeline for this journal, I would greatly appreciate any advice you might have on how to proceed.

Thank you for your time.
