+++ src/elements/triangle/internalForcesShellTriangle.m
@@ -26,7 +26,8 @@ function [fs, ks, fintLocCoord, rotMat] = internalForcesShellTriangle(elemCoords
   flag_first_mod  = 0; % 1 if battini modification is used - 0 if not / local rotations
   flag_second_mod = 0;  % 1 if battini modification is used - 0 if not / out of plane disps = 0
   flag_third_mod  = 0;  % 1 if battini modification is used - 0 if not / quaternions
-  flag_OPT        = 1;  % 1 if optimal triangle membrane element is used
+  % flag_OPT        = 1;  % 1 if optimal triangle membrane element is used
+  global flag_OPT
 
   % material and geometric parameters
   young_modulus = modelParams(1);
@@ -158,7 +159,7 @@ function [fs, ks, fintLocCoord, rotMat] = internalForcesShellTriangle(elemCoords
   pl_full(index_full) = pl;
 
   % calculating the linear stiffness matrix and internal force vector of the shell element in local coordinates
-  [Kl_full, fintLocCoord] = localShellTriangle(x02, x03, y03, young_modulus, poisson_ratio, h, pl_full, flag_OPT, r1_g, r2_g, r3_g, To);
+  [Kl_full, fintLocCoord] = localShellTriangle(x02, x03, y03, young_modulus, poisson_ratio, h, pl_full, flag_OPT);
 
   % Reduces Kl matrix to number of dofs considered
   Kl = Kl_full(index_full, index_full);
