+++ src/mesh/elem2NodalLoads.m
@@ -78,23 +78,41 @@ function fext = elem2NodalLoads (Conec, indBC, elements, boundaryCond, Nodes)
       factor = lengthElem * thickness * 0.5;
 
       if strcmp(loadCoordSys, 'global')
+        % only for loads on x axis and sides parallel to XY or XZ
         qx = loadvals(1) ;
-        qy = loadvals(2) ;
+        % qy = loadvals(2) ;
+        % qz = loadvals(3) ;
         xaxis = [1,0,0];
+        yaxis = [0,1,0];
+        zaxis = [0,0,1];
         
         normalVector = cross(xaxis,directionVector);
         nvector = normalVector / norm(normalVector);
+        
         if norm(normalVector) == 0 
           angle = 0;
         else
           angle = asin(normalVector/( norm(xaxis)*norm(directionVector)*nvector )); % cross(a,b) = |a|*|b|*sin(theta)*n
         end
-        q_perp = qx*sin(angle) - qy*cos(angle);
-                
-        % Aedge = lengthElem * thickness; % edge area
-        Mz_q = q_perp*thickness*lengthElem^2/12; % 
-        % Mz_q = qx*thickness*lengthElem^2/8; % 
-        % Mz_q=0;
+        q_perp = qx*sin(angle);
+
+        % Projection of normal vector into x,y,z axes
+        % indBC
+        % directionVector
+        % nvector
+
+        Mq = q_perp*thickness*lengthElem^2/8; %
+        Mq=0;
+        cos_nx = abs(xaxis * nvector');
+        cos_ny = abs(yaxis * nvector');
+        cos_nz = abs(zaxis * nvector');
+        
+        Mqx = Mq*cos_nx;
+        Mqy = Mq*cos_ny;
+        Mqz = Mq*cos_nz;
+
+        Mq_opt = [0 Mqx 0 -Mqy 0 Mqz; 0 Mqx 0 Mqy 0 -Mqz] ;
+         
         Fx = loadvals(1) * factor;
         Mx = loadvals(2) * factor;
         Fy = loadvals(3) * factor;
@@ -102,8 +120,8 @@ function fext = elem2NodalLoads (Conec, indBC, elements, boundaryCond, Nodes)
         Fz = loadvals(5) * factor;
         Mz = loadvals(6) * factor;
         
-        elemNodeLoadsMatrix = [Fx Mx Fy My Fz Mz+Mz_q; Fx Mx Fy My Fz Mz-Mz_q] ;
-        % elemNodeLoadsMatrix
+        elemNodeLoadsMatrix = [Fx Mx Fy My Fz Mz; Fx Mx Fy My Fz Mz] + Mq_opt;
+        
       elseif strcmp(loadCoordSys, 'local')
         % consider a 90 degrees rotation of the oriented vector of the line element
         % tangent unitary vector
