### 2025-09-17 f6f192f
varias
