+++ src/elements/frame/frame_internal_force.m
@@ -1,195 +1,0 @@
-% Copyright 2024, ONSAS Authors (see documentation)
-%
-% This file is part of ONSAS.
-%
-% ONSAS is free software: you can redistribute it and/or modify
-% it under the terms of the GNU General Public License as published by
-% the Free Software Foundation, either version 3 of the License, or
-% (at your option) any later version.
-%
-% ONSAS is distributed in the hope that it will be useful,
-% but WITHOUT ANY WARRANTY; without even the implied warranty of
-% MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-% GNU General Public License for more details.
-%
-% You should have received a copy of the GNU General Public License
-% along with ONSAS.  If not, see <https://www.gnu.org/licenses/>.
-%
-
-function  [ fs, ks, stress, rotData, localInternalForces ] = frame_internal_force( ...
-  elemCoords, elemCrossSecParams, elemConstitutiveParams, Ue ) ;
-  
-  % element coordinates
-  xs = elemCoords(:) ;
-
-  % ----- material and geometric params ------
-  E   = elemConstitutiveParams(2) ;
-  nu  = elemConstitutiveParams(3) ;
-  G   = E/(2*(1+nu))              ;
-  % ----- extract cross section properties ---
-  [Area, J, Iyy, Izz, ~] = crossSectionProps ( elemCrossSecParams, 0 ) ; % select a ficticious elemrho 
-  % ------------------------------------------
-
-  % compute corotational matrices rotation 
-  [R0, Rr, Rg1, Rg2, Rroof1, Rroof2] = corotRotMatrices( Ue, elemCoords ) ;
-
-  % --- global displacements ---
-  % permut indexes according to Battini's nomenclature
-  dg = switchToBattiniNom( Ue ) ;
-  % global thetas
-  tg1 = dg(  4:6  ) ;
-  tg2 = dg( 10:12 ) ;
-  % -------------------------------
-
-  % length and coords of the element  
-  [x21, d21, l, l0] = corotLenCoords(xs ,dg) ;
-
-  % --- local displacements ---
-
-  % axial displacement
-  u   = l - l0 ;
-
-  %%%%%%%%%%%%%%%%%%%%%%%%%%%
-  % temporary
-  %%%%%%%%%%%%%%%%%%%%%%%%%%%
-  global pretension_strain
-  if ~isempty( pretension_strain )
-    % axial displacement
-    u   = u + pretension_strain * l0;
-  end
-  %%%%%%%%%%%%%%%%%%%%%%%%%%%
-  
-  % local rotations
-  tl1 = logar( Rroof1 ) ;
-  tl2 = logar( Rroof2 ) ;  
-  locDisp = [ u tl1' tl2' ] ;
-  % -------------------------------
-
-  % --- auxiliary vector and matrices  ---
-  % aux zero matrices 
-  [I3, O3, O1, II] = corotZeros() ;
-
-  % auxiliary q base 
-  q1g = Rg1 * R0 * [0 1 0]' ;
-  q2g = Rg2 * R0 * [0 1 0]' ;
-  qg  = ( q1g + q2g ) / 2   ;
-
[…]
