+++ src/interface/initONSAS.m
@@ -18,198 +18,175 @@
 
 function [modelCurrSol, modelProperties, BCsData] = initONSAS(materials, elements, boundaryConds, initialConds, mesh, analysisSettings, otherParams)
 
-  % checks if the fields defined are correct or not
-  checkONSASFields(materials, elements, boundaryConds, initialConds, mesh, analysisSettings, otherParams);
 
-  % md set defaults
-  [materials, elements, boundaryConds, analysisSettings, otherParams] = setDefaults(materials, elements, boundaryConds, analysisSettings, otherParams);
+global TZERO
 
-  % md sets the current version and welcomes user
-  ONSASversion = '0.4.0';
-  welcomeMessage(ONSASversion, otherParams);
+% checks if the fields defined are correct or not
+checkONSASFields(materials, elements, boundaryConds, initialConds, mesh, analysisSettings, otherParams )
 
-  % creates outputdir in current location
-  createOutputDir(otherParams);
+%md set defaults
+[ materials, elements, boundaryConds, analysisSettings, otherParams ] = setDefaults( materials, elements, boundaryConds, analysisSettings, otherParams ) ;
 
-  % =================================================================
-  % md process boundary conds information and construct BCsData struct
-  [Conec, Nodes, factorLoadsFextCell, loadFactorsFuncCell, diriDofs, neumDofs, KS, userLoadsFilename] ...
-    = boundaryCondsProcessing(mesh, materials, elements, boundaryConds, analysisSettings);
+%md sets the current version and welcomes user
+ONSASversion = '0.3.0'  ; welcome_message( ONSASversion, otherParams );
 
-  BCsData = constructBCsData(factorLoadsFextCell, loadFactorsFuncCell, neumDofs, KS, userLoadsFilename);
-  % =================================================================
+% creates outputdir in current location
+create_outputDir( otherParams )
 
-  nTimes = round(analysisSettings.finalTime / analysisSettings.deltaT) + 1; % number of times (including t=0)
-  % if length( otherParams.plotParamsVector ) > 1
-  %   nplots = min( [ nTimes otherParams.plotParamsVector(2) ] ) ;
-  % else
-  %   % default value: all
-  nplots = nTimes;
-  % end
-  timesPlotsVec = round(linspace(1, nTimes, nplots)');
+% =================================================================
+%md process boundary conds information and construct BCsData struct
+[ Conec, Nodes, factorLoadsFextCell, loadFactorsFuncCell, diriDofs, neumDofs, KS, userLoadsFilename ] ...
+  = boundaryCondsProcessing( mesh, materials, elements, boundaryConds, analysisSettings ) ;
 
-  % =================================================================
-  % md construct modelProperties struct
-  modelProperties = constructModelProperties(Nodes, Conec, materials, elements, analysisSettings, otherParams, timesPlotsVec);
-  % =================================================================
+BCsData = construct_BCsData( factorLoadsFextCell, loadFactorsFuncCell, neumDofs, KS, userLoadsFilename );
+% =================================================================
 
-  %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-  % TEMPORARY
-  %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-  global spitMatrices
-  if spitMatrices == true
-    save('-mat', 'output/loads.mat', 'factorLoadsFextCell');
+
[…]
