+++ src/interface/solveONSAS.m
@@ -21,136 +21,144 @@
 %
 function [matUs, loadFactorsMat, modelSolutions] = solveONSAS(modelCurrSol, modelProperties, BCsData)
 
-  % initialize structures to store solutions
-  matUs          = modelCurrSol.U;
-  loadFactorsMat = modelCurrSol.currLoadFactorsVals;
-  matUdots       = modelCurrSol.Udot;
-  cellFint       = {modelCurrSol.localInternalForces};
+global historic_parameters
+global Timex
 
-  modelSolutions = {modelCurrSol};
-  cellStress     = {modelCurrSol.Stress}; % cell with a matrix with all elements stresses at each time
+% initialize structures to store solutions
+matUs          = modelCurrSol.U                   ;
+loadFactorsMat = modelCurrSol.currLoadFactorsVals ;
+matUdots       = modelCurrSol.Udot                ;
+cellFint       = {modelCurrSol.localInternalForces} ; 
 
-  % Incremental time analysis
-  % sets stopping boolean to false
-  finalTimeReachedBoolean = false;
-  % and starts the iteration
-  fprintf('|                                                 |\n');
-  fprintf('| Analysis progress:   |0       50       100| %%   |\n');
-  fprintf('|                      |');
+modelSolutions = {modelCurrSol};
+cellStress     = {modelCurrSol.Stress}  ; % cell with a matrix with all elements stresses at each time
 
-  % iteration variables
-  iterations_average = 0;
-  iterations_maximum = 0;
-  iterations_strop_crit_vec = [0 0 0];
+% Incremental time analysis
+% sets stopping boolean to false
+finalTimeReachedBoolean = false ;
+% and starts the iteration
+fprintf('|                                                 |\n')
+fprintf('| Analysis progress:   |0       50       100| %%   |\n')
+fprintf('|                      |')
 
-  % progress bar variables
-  plotted_bars = 0;
-  aux_time = cputime();
+% iteration variables
+iterations_average = 0 ;  iterations_maximum = 0 ;
+iterations_strop_crit_vec = [ 0 0 0 ] ;
 
-  while finalTimeReachedBoolean == false
-    plotted_bars = progressBarPlot(modelCurrSol, modelProperties, plotted_bars);
+% progress bar variables
+plotted_bars = 0 ; aux_time = cputime() ;
 
-    % compute the model state at next time
-    modelNextSol = timeStepIteration(modelCurrSol, modelProperties, BCsData);
+while finalTimeReachedBoolean == false
 
-    % iterations average
-    iterations_average = ...
-      (iterations_average * (modelNextSol.timeIndex - 2) + modelNextSol.timeStepIters) / ...
-      (modelNextSol.timeIndex - 1);
+  % disp(' ====================================================') ;
+  % disp(' /\  /\  /\  /\  /\  NEW TIME STEP /\  /\  /\  /\  /\') ;
+  plotted_bars = progressBarPlot( modelCurrSol, modelProperties, plotted_bars) ;
 
-    % iterations max
-    iterations_maximum = max(iterations_maximum, modelNextSol.timeStepIters);
-    iterations_strop_crit_vec(modelNextSol.timeStepStopCrit) = ...
-      iterations_strop_crit_vec(modelNextSol.timeStepStopCrit) + 1;
+  Timex = modelCurrSol.currTime ;
 
-    % check if final time was reached
[…]
