+++ src/interface/ONSAS_solve.m
@@ -1,164 +1,0 @@
-% Copyright 2024, ONSAS Authors (see documentation)
-%
-% This file is part of ONSAS.
-%
-% ONSAS is free software: you can redistribute it and/or modify
-% it under the terms of the GNU General Public License as published by
-% the Free Software Foundation, either version 3 of the License, or
-% (at your option) any later version.
-%
-% ONSAS is distributed in the hope that it will be useful,
-% but WITHOUT ANY WARRANTY; without even the implied warranty of
-% MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-% GNU General Public License for more details.
-%
-% You should have received a copy of the GNU General Public License
-% along with ONSAS.  If not, see <https://www.gnu.org/licenses/>.
-%
-
-%
-% Function that performs the time analysis with the model structs as input.
-%
-function [ matUs, loadFactorsMat, modelSolutions ] = ONSAS_solve( modelCurrSol, modelProperties, BCsData )
-
-global historic_parameters
-global Timex
-
-% initialize structures to store solutions
-matUs          = modelCurrSol.U                   ;
-loadFactorsMat = modelCurrSol.currLoadFactorsVals ;
-matUdots       = modelCurrSol.Udot                ;
-cellFint       = {modelCurrSol.localInternalForces} ; 
-
-modelSolutions = {modelCurrSol};
-cellStress     = {modelCurrSol.Stress}  ; % cell with a matrix with all elements stresses at each time
-
-% Incremental time analysis
-% sets stopping boolean to false
-finalTimeReachedBoolean = false ;
-% and starts the iteration
-fprintf('|                                                 |\n')
-fprintf('| Analysis progress:   |0       50       100| %%   |\n')
-fprintf('|                      |')
-
-% iteration variables
-iterations_average = 0 ;  iterations_maximum = 0 ;
-iterations_strop_crit_vec = [ 0 0 0 ] ;
-
-% progress bar variables
-plotted_bars = 0 ; aux_time = cputime() ;
-
-while finalTimeReachedBoolean == false
-
-  % disp(' ====================================================') ;
-  % disp(' /\  /\  /\  /\  /\  NEW TIME STEP /\  /\  /\  /\  /\') ;
-  plotted_bars = progressBarPlot( modelCurrSol, modelProperties, plotted_bars) ;
-
-  Timex = modelCurrSol.currTime ;
-
-  % compute the model state at next time
-  modelNextSol = timeStepIteration( modelCurrSol, modelProperties, BCsData ) ;
-
-  historic_parameters = [ historic_parameters ;  modelNextSol.previousStateCell ] ;
-
-  % iterations average
-  iterations_average = ...
-    (   iterations_average* (modelNextSol.timeIndex-2) + modelNextSol.timeStepIters ) ...
-    / ( modelNextSol.timeIndex-1) ;
-
-  % iterations max
-  iterations_maximum = max( iterations_maximum, modelNextSol.timeStepIters ) ;
-  iterations_strop_crit_vec( modelNextSol.timeStepStopCrit ) = ...
-    iterations_strop_crit_vec( modelNextSol.timeStepStopCrit ) + 1 ;
-
-  % check if final time was reached
-  finalTimeReachedBoolean = ( modelNextSol.currTime - modelProperties.analysisSettings.finalTime ) ...
[…]
