+++ src/interface/initONSAS.m
@@ -18,175 +18,200 @@
 
 function [modelCurrSol, modelProperties, BCsData] = initONSAS(materials, elements, boundaryConds, initialConds, mesh, analysisSettings, otherParams)
 
+  global TZERO
 
-global TZERO
+  % checks if the fields defined are correct or not
+  checkONSASFields(materials, elements, boundaryConds, initialConds, mesh, analysisSettings, otherParams);
 
-% checks if the fields defined are correct or not
-checkONSASFields(materials, elements, boundaryConds, initialConds, mesh, analysisSettings, otherParams )
+  % md set defaults
+  [materials, elements, boundaryConds, analysisSettings, otherParams] = setDefaults(materials, elements, boundaryConds, analysisSettings, otherParams);
 
-%md set defaults
-[ materials, elements, boundaryConds, analysisSettings, otherParams ] = setDefaults( materials, elements, boundaryConds, analysisSettings, otherParams ) ;
+  % md sets the current version and welcomes user
+  ONSASversion = '0.4.0';
+  welcomeMessage(ONSASversion, otherParams);
 
-%md sets the current version and welcomes user
-ONSASversion = '0.3.0'  ; welcome_message( ONSASversion, otherParams );
+  % creates outputdir in current location
+  createOutputDir(otherParams);
 
-% creates outputdir in current location
-create_outputDir( otherParams )
+  % =================================================================
+  % md process boundary conds information and construct BCsData struct
+  [Conec, Nodes, factorLoadsFextCell, loadFactorsFuncCell, diriDofs, neumDofs, KS, userLoadsFilename] ...
+    = boundaryCondsProcessing(mesh, materials, elements, boundaryConds, analysisSettings);
 
-% =================================================================
-%md process boundary conds information and construct BCsData struct
-[ Conec, Nodes, factorLoadsFextCell, loadFactorsFuncCell, diriDofs, neumDofs, KS, userLoadsFilename ] ...
-  = boundaryCondsProcessing( mesh, materials, elements, boundaryConds, analysisSettings ) ;
+  BCsData = constructBCsData(factorLoadsFextCell, loadFactorsFuncCell, neumDofs, KS, userLoadsFilename);
+  % =================================================================
 
-BCsData = construct_BCsData( factorLoadsFextCell, loadFactorsFuncCell, neumDofs, KS, userLoadsFilename );
-% =================================================================
+  nTimes = round(analysisSettings.finalTime / analysisSettings.deltaT) + 1; % number of times (including t=0)
+  % if length( otherParams.plotParamsVector ) > 1
+  %   nplots = min( [ nTimes otherParams.plotParamsVector(2) ] ) ;
+  % else
+  %   % default value: all
+  nplots = nTimes;
+  % end
+  timesPlotsVec = round(linspace(1, nTimes, nplots)');
 
+  % =================================================================
+  % md construct modelProperties struct
+  modelProperties = constructModelProperties(Nodes, Conec, materials, elements, analysisSettings, otherParams, timesPlotsVec);
+  % =================================================================
 
-nTimes = round( analysisSettings.finalTime / analysisSettings.deltaT ) + 1 ; % number of times (including t=0)
-% if length( otherParams.plotParamsVector ) > 1
-%   nplots = min( [ nTimes otherParams.plotParamsVector(2) ] ) ;
-% else
-%   % default value: all
[…]
