+++ src/elements/frame/Versiones Anteriores/framePlastic.m
@@ -14,7 +14,7 @@
 %
 % You should have received a copy of the GNU General Public License
 % along with ONSAS.  If not, see <https://www.gnu.org/licenses/>.
- 
+
 % =========================================================================
 
 % Euler-Bernoulli element with embeded discontinuity
@@ -29,123 +29,121 @@
 
 % =========================================================================
 
-function [ soft_hinge_boolean, Fint, M1, Kelement, kpn1, xin11, xin21, alfan1, xd, xdi, tM] = framePlastic(soft_hinge_boolean, dnk, kpn, xin1, xin2, alfan, xd, xdi, tM, elemParams, elastoplasticParams)
+function [soft_hinge_boolean, Fint, M1, Kelement, kpn1, xin11, xin21, alfan1, xd, xdi, tM] = framePlastic(soft_hinge_boolean, dnk, kpn, xin1, xin2, alfan, xd, xdi, tM, elemParams, elastoplasticParams)
 
+  % --- element params ---
+  l  = elemParams(1);
+  A  = elemParams(2);
+  Iy = elemParams(3);
 
-% --- element params ---
-l  = elemParams(1) ;
-A  = elemParams(2) ;
-Iy = elemParams(3) ;
+  % --- elastoplastic params ---
+  E   = elastoplasticParams(1);
+  Mc  = elastoplasticParams(2);
+  My  = elastoplasticParams(3);
+  Mu  = elastoplasticParams(4);
+  kh1 = elastoplasticParams(5);
+  kh2 = elastoplasticParams(6);
+  Ks  = elastoplasticParams(7);
 
-% --- elastoplastic params ---
-E   = elastoplasticParams(1) ;
-Mc  = elastoplasticParams(2) ;
-My  = elastoplasticParams(3) ;
-Mu  = elastoplasticParams(4) ;
-kh1 = elastoplasticParams(5) ;
-kh2 = elastoplasticParams(6) ;
-Ks  = elastoplasticParams(7) ;
+  uvector     = dnk(1:2);
+  vvector     = dnk(3:4);
+  thetavector = dnk(5:6);
 
-uvector     = dnk(1:2) ;
-vvector     = dnk(3:4) ;
-thetavector = dnk(5:6) ;
+  Kfd    = zeros(6, 6);
+  Kfalfa = zeros(6, 6);
+  Khd    = zeros(6, 6);
+  Khalfa = 0;
 
-Kfd    = zeros(6,6) ;
-Kfalfa = zeros(6,6) ;
-Khd    = zeros(6,6) ;
-Khalfa = 0 ;
+  Cep    = 0;
+  Fint   = 0;
 
-Cep    = 0 ;
-Fint   = 0 ;
+  % Gauss-Lobatto Quadrature with 3 integration points [a (a+b)/2 b]
+  npi = 3;
+  xpi = [0 l / 2 l];
+  wpi = [1 / 3 4 / 3 1 / 3] * l * 0.5;
 
-% Gauss-Lobatto Quadrature with 3 integration points [a (a+b)/2 b]
-npi = 3 ;
-xpi = [0 l/2 l] ;
-wpi = [1/3 4/3 1/3]*l*0.5 ;
+  % integration (Gauss-Lobatto)
+  % and calculation of values of internal parameters at integration points
 
-% integration (Gauss-Lobatto)
-% and calculation of values of internal parameters at integration points
+  % initial values of bulk moments
+  M1 = zeros(npi, 1);
 
-% initial values of bulk moments
-M1 = zeros(npi,1) ;
[…]
