+++ src/core/system_assembler.m
@@ -15,154 +15,158 @@
 % You should have received a copy of the GNU General Public License
 % along with ONSAS.  If not, see <https://www.gnu.org/licenses/>.
 %
-function [systemDeltauMatrix, systemDeltauRHS, FextG, fs, nexTimeLoadFactors, fnorms, exportFirstMatrices ] = system_assembler( modelProperties, BCsData, Ut, Udott, Udotdott, Utp1, Udottp1, Udotdottp1, nextTime, nexTimeLoadFactors, previousStateCell )
+function [systemDeltauMatrix, systemDeltauRHS, FextG, fs, nexTimeLoadFactors, fnorms, exportFirstMatrices] = system_assembler(modelProperties, BCsData, Ut, Udott, Udotdott, Utp1, Udottp1, Udotdottp1, nextTime, nexTimeLoadFactors, previousStateCell)
 
-  analysisSettings = modelProperties.analysisSettings ;
-  nodalDispDamping = modelProperties.nodalDispDamping ;
-  neumdofs = BCsData.neumDofs ;
-  
-  [fs, ~, mats, ~ ] = assembler( modelProperties.Conec, modelProperties.elements, modelProperties.Nodes, modelProperties.materials, BCsData(1).KS, Utp1, Udottp1, Udotdottp1, analysisSettings, [1 0 1 0], nodalDispDamping, nextTime, previousStateCell  ) ;
+  analysisSettings = modelProperties.analysisSettings;
+  nodalDispDamping = modelProperties.nodalDispDamping;
+  neumdofs = BCsData.neumDofs;
 
-  Fint = fs{1} ;  Fvis =  fs{2} ;  Fmas = fs{3} ; Faero = fs{4} ;
-  
+  [fs, ~, mats, ~] = assembler(modelProperties.Conec, modelProperties.elements, modelProperties.Nodes, modelProperties.materials, BCsData(1).KS, Utp1, Udottp1, Udotdottp1, analysisSettings, [1 0 1 0], nodalDispDamping, nextTime, previousStateCell);
 
-  exportFirstMatrices = false ;
-  KT   = mats{1} ; 
+  Fint = fs{1};
+  Fvis =  fs{2};
+  Fmas = fs{3};
+  Faero = fs{4};
 
-  if strcmp( analysisSettings.methodName, 'newmark' ) || strcmp( analysisSettings.methodName, 'alphaHHT' )
-    dampingMat = mats{2} ;
-    massMat    = mats{3} ;
+  exportFirstMatrices = false;
+  KT   = mats{1};
+
+  if strcmp(analysisSettings.methodName, 'newmark') || strcmp(analysisSettings.methodName, 'alphaHHT')
+    dampingMat = mats{2};
+    massMat    = mats{3};
 
     if modelProperties.exportFirstMatrices
-      KTred      = KT( neumdofs, neumdofs );
-      massMatred = massMat(neumdofs,neumdofs);
-      save('-mat', 'output/matrices.mat', 'KT','massMat','neumdofs' );
-      figure
-      spy(full(KT)), title('stiffness')
-      figure
-      spy(full(massMat)), title('mass')
-      fprintf('matrices exported.\n--------\n')
+      KTred      = KT(neumdofs, neumdofs);
+      massMatred = massMat(neumdofs, neumdofs);
+      save('-mat', 'output/matrices.mat', 'KT', 'massMat', 'neumdofs');
+      figure;
+      spy(full(KT));
+      title('stiffness');
+      figure;
+      spy(full(massMat));
+      title('mass');
+      fprintf('matrices exported.\n--------\n');
     end
   end
 
[…]

[…]
