+++ src/elements/frame/frame2DPlasticInternalForce.m
@@ -27,10 +27,10 @@
 % Adnan Ibrahimbegović / Ecole normale supérieure de Cachan
 % =========================================================================
 
-function [fs, ks, fintLocCoord, params_plastic_2Dframe_np1] = frame2D_plastic_internal_force(elemNodesxyzRefCoords, ...
-                                                                                               elemCrossSecParams, ...
-                                                                                               modelParams, ...
-                                                                                               elemDisps, params_plastic_2Dframe, element)
+function [fs, ks, fintLocCoord, params_plastic_2Dframe_np1] = frame2DPlasticInternalForce(elemNodesxyzRefCoords, ...
+                                                                                          elemCrossSecParams, ...
+                                                                                          modelParams, ...
+                                                                                          elemDisps, params_plastic_2Dframe, element)
 
   % global soft_activation
   global Timex
