+++ src/core/assembler.m
@@ -215,9 +215,9 @@ function [fsCell, stressMat, tangMatsCell, localInternalForces, strain_vec, acum
         params_plastic_2Dframe(elem, :) = previousStateCell(elem, :);
 
         [fs, ks, fintLocCoord, aux] = frame2D_plastic_internal_force(elemNodesxyzRefCoords, ...
-                                                                       elemCrossSecParams, ...
-                                                                       modelParams, ...
-                                                                       elemDisps, params_plastic_2Dframe(elem, :), elem);
+                                                                     elemCrossSecParams, ...
+                                                                     modelParams, ...
+                                                                     elemDisps, params_plastic_2Dframe(elem, :), elem);
 
         Nx = 0;
         My = 0;
@@ -236,9 +236,9 @@ function [fsCell, stressMat, tangMatsCell, localInternalForces, strain_vec, acum
 
         if dynamicProblemBool
           [fs, ks] = frame_inertial_force(elemNodesxyzRefCoords, elemCrossSecParams, ...
-                                             [1 modelParams], elemDisps, ...
-                                             dotdispsElem, dotdotdispsElem, ...
-                                             density, massMatType, analysisSettings);
+                                          [1 modelParams], elemDisps, ...
+                                          dotdispsElem, dotdotdispsElem, ...
+                                          density, massMatType, analysisSettings);
 
           Fmase = fs{3};
           Ce = ks{2};
