+++ src/core/system_assembler.m
@@ -1,172 +1,0 @@
-% Copyright 2024, ONSAS Authors (see documentation)
-%
-% This file is part of ONSAS.
-%
-% ONSAS is free software: you can redistribute it and/or modify
-% it under the terms of the GNU General Public License as published by
-% the Free Software Foundation, either version 3 of the License, or
-% (at your option) any later version.
-%
-% ONSAS is distributed in the hope that it will be useful,
-% but WITHOUT ANY WARRANTY; without even the implied warranty of
-% MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-% GNU General Public License for more details.
-%
-% You should have received a copy of the GNU General Public License
-% along with ONSAS.  If not, see <https://www.gnu.org/licenses/>.
-%
-function [systemDeltauMatrix, systemDeltauRHS, FextG, fs, nexTimeLoadFactors, fnorms, exportFirstMatrices] = system_assembler(modelProperties, BCsData, Ut, Udott, Udotdott, Utp1, Udottp1, Udotdottp1, nextTime, nexTimeLoadFactors, previousStateCell)
-
-  analysisSettings = modelProperties.analysisSettings;
-  nodalDispDamping = modelProperties.nodalDispDamping;
-  neumdofs = BCsData.neumDofs;
-
-  [fs, ~, mats, ~] = assembler(modelProperties.Conec, modelProperties.elements, modelProperties.Nodes, modelProperties.materials, BCsData(1).KS, Utp1, Udottp1, Udotdottp1, analysisSettings, [1 0 1 0], nodalDispDamping, nextTime, previousStateCell);
-
-  Fint = fs{1};
-  Fvis =  fs{2};
-  Fmas = fs{3};
-  Faero = fs{4};
-
-  exportFirstMatrices = false;
-  KT   = mats{1};
-
-  if strcmp(analysisSettings.methodName, 'newmark') || strcmp(analysisSettings.methodName, 'alphaHHT')
-    dampingMat = mats{2};
-    massMat    = mats{3};
-
-    if modelProperties.exportFirstMatrices
-      KTred      = KT(neumdofs, neumdofs);
-      massMatred = massMat(neumdofs, neumdofs);
-      save('-mat', 'output/matrices.mat', 'KT', 'massMat', 'neumdofs');
-      figure;
-      spy(full(KT));
-      title('stiffness');
-      figure;
-      spy(full(massMat));
-      title('mass');
-      fprintf('matrices exported.\n--------\n');
-    end
-  end
-
-  if strcmp(modelProperties.analysisSettings.methodName, 'newtonRaphson')
-
-    [FextG, nexTimeLoadFactors]  = computeFext(modelProperties, BCsData, nextTime, length(Fint), [],  {Utp1});
-
-    rhat      =   Fint (BCsData.neumDofs) ...
-                - FextG(BCsData.neumDofs) ...
-                - Faero(BCsData.neumDofs);
-
-    systemDeltauRHS = -rhat;
-
-    systemDeltauMatrix = KT (neumdofs, neumdofs);
-
-    % -----------------------------------------------------------------------------------
-    % newmark
-    % -----------------------------------------------------------------------------------
-  elseif strcmp(modelProperties.analysisSettings.methodName, 'arcLength')
-
-    [FextG, nexTimeLoadFactors]  = computeFext(modelProperties, BCsData, nextTime, length(Fint), nexTimeLoadFactors, {Utp1});
-
-    foundLoadCase = false;
-    loadCase = 1;
-    while ~foundLoadCase
-      if isempty(BCsData.factorLoadsFextCell{loadCase})
-        loadCase = loadCase + 1;
-      else
-        foundLoadCase = true;
-      end
-    end
-
[…]
