### 2025-11-20 455bb0f
integrando
