+++ src/interface/ONSAS_init.m
@@ -1,189 +1,0 @@
-% Copyright 2024, ONSAS Authors (see documentation)
-%
-% This file is part of ONSAS.
-%
-% ONSAS is free software: you can redistribute it and/or modify
-% it under the terms of the GNU General Public License as published by
-% the Free Software Foundation, either version 3 of the License, or
-% (at your option) any later version.
-%
-% ONSAS is distributed in the hope that it will be useful,
-% but WITHOUT ANY WARRANTY; without even the implied warranty of
-% MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-% GNU General Public License for more details.
-%
-% You should have received a copy of the GNU General Public License
-% along with ONSAS.  If not, see <https://www.gnu.org/licenses/>.
-%
-
-function [modelCurrSol, modelProperties, BCsData] = ONSAS_init(materials, elements, boundaryConds, initialConds, mesh, analysisSettings, otherParams)
-
-  global TZERO
-
-  % checks if the fields defined are correct or not
-  checkONSASFields(materials, elements, boundaryConds, initialConds, mesh, analysisSettings, otherParams);
-
-  % md set defaults
-  [materials, elements, boundaryConds, analysisSettings, otherParams] = setDefaults(materials, elements, boundaryConds, analysisSettings, otherParams);
-
-  % md sets the current version and welcomes user
-  ONSASversion = '0.3.0';
-  welcome_message(ONSASversion, otherParams);
-
-  % creates outputdir in current location
-  create_outputDir(otherParams);
-
-  % =================================================================
-  % md process boundary conds information and construct BCsData struct
-  [Conec, Nodes, factorLoadsFextCell, loadFactorsFuncCell, diriDofs, neumDofs, KS, userLoadsFilename] ...
-    = boundaryCondsProcessing(mesh, materials, elements, boundaryConds, analysisSettings);
-
-  BCsData = construct_BCsData(factorLoadsFextCell, loadFactorsFuncCell, neumDofs, KS, userLoadsFilename);
-  % =================================================================
-
-  nTimes = round(analysisSettings.finalTime / analysisSettings.deltaT) + 1; % number of times (including t=0)
-  % if length( otherParams.plotParamsVector ) > 1
-  %   nplots = min( [ nTimes otherParams.plotParamsVector(2) ] ) ;
-  % else
-  %   % default value: all
-  nplots = nTimes;
-  % end
-  timesPlotsVec = round(linspace(1, nTimes, nplots)');
-
-  % =================================================================
-  % md construct modelProperties struct
-  modelProperties = construct_modelProperties(Nodes, Conec, materials, elements, analysisSettings, otherParams, timesPlotsVec);
-  % =================================================================
-
-  %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-  %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-  global spitMatrices
-  if spitMatrices == true
-    save('-mat', 'output/loads.mat', 'factorLoadsFextCell');
-  end
-  %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-  %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-
-  % =================================================================
-  % md process initial conds and construct modelSol struct
-  currTime  = 0;
-  timeIndex = 1;
-
-  timeStepIters    = 0;
-  timeStepStopCrit = 0;
-
-  % process initial conditions
-  nNodes = size(mesh.nodesCoords, 1);
[…]
