+++ examples/plasticSimplySupported/plasticSimplySupported.m
@@ -1,197 +1,0 @@
-% Copyright 2024, ONSAS Authors (see documentation)
-%
-% This file is part of ONSAS.
-%
-% ONSAS is free software: you can redistribute it and/or modify
-% it under the terms of the GNU General Public License as published by
-% the Free Software Foundation, either version 3 of the License, or
-% (at your option) any later version.
-%
-% ONSAS is distributed in the hope that it will be useful,
-% but WITHOUT ANY WARRANTY; without even the implied warranty of
-% MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-% GNU General Public License for more details.
-%
-% You should have received a copy of the GNU General Public License
-% along with ONSAS.  If not, see <https://www.gnu.org/licenses/>.
-
-% =========================================================================
-
-close all;
-clear;
-addpath(genpath([pwd '/../../src']));
-
-% assumed XY plane
-
-% -------------------------------------------
-% scalar parameters
-% material
-EI  = 77650;       % KN.m^2
-kh1 = 29400;       % KN.m^2
-kh2 = 273;
-Ks  = -18000;      % KN.m
-
-nu = 0.3;          % Poisson's ratio
-
-% geometry
-l1  = 1.0;              % m
-l2  = 1.0;              % m
-ty = 0.3;              % width cross section
-tz = 0.4;              % height cross section
-Inertia = tz * ty^3 / 12;  % m^4
-
-E = EI / Inertia;        % KN/m^2 [KPa]
-
-A  = ty * tz;            % m^2
-Mc = 37.9;             % KN.m
-My = 268;
-Mu = 374;
-
-materials             = struct();
-materials.modelName   = 'plastic-2Dframe';
-materials.modelParams = [E Mc My Mu kh1 kh2 Ks nu];
-
-Pc = Mc * (l1 + l2) / (l1 * l2);
-
-deltac = Pc * (l2 * ((l1 + l2)^2 - l2^2)^(1.5)) / (9 * sqrt(3) * EI);
-
-elements             = struct();
-elements(1).elemType = 'node';
-
-elements(2).elemType = 'frame';
-elements(2).elemCrossSecParams = {'generic'; [A 1 Inertia Inertia] };
-
-boundaryConds                  = {};
-boundaryConds(1).imposDispDofs = [1 2 3 4 5];
-boundaryConds(1).imposDispVals = [0 0 0 0 0];
-
-boundaryConds(2).loadsCoordSys = 'global';
-boundaryConds(2).loadsBaseVals = [0 0 -1 0 0 0];
-boundaryConds(2).loadsTimeFact = @(t) t;
-boundaryConds(2).imposDispDofs = [2 4 5];
-boundaryConds(2).imposDispVals = [0 0 0];
-
-% The coordinates of the nodes of the mesh are given by the matrix:
-mesh = {};
-num_elem = 2;
-xs = linspace(0, l1 + l2, num_elem + 1);
-mesh.nodesCoords = [0 0 0;  l1 0 0; l1 + l2 0 0];
-
-mesh.conecCell = {};
-
-mesh.conecCell{ 1, 1 } = [0 1 1  1]; % node with fixed end support
-mesh.conecCell{ end + 1, 1 } = [0 1 1  3]; % node with fixed end support
-mesh.conecCell{ end + 1, 1 } = [0 1 2  2]; % loaded node
-
-for k = 1:num_elem
[…]
