+++ miss_hit.cfg
@@ -27,7 +27,7 @@ metric "cnest": limit 6
 metric "file_length": limit 1000
 # "cyc" (Cyclomatic Complexity): Maximum allowed complexity of a single function
 # A lower value ensures that functions remain simple and focused.
-metric "cyc": limit 53
+metric "cyc": limit 55
 # "parameters": Maximum allowed number of parameters for a single function
 # Encourages simpler function signatures and improves readability.
 metric "parameters": limit 29
