+++ examples/plasticCantileverBeam/plasticCantileverBeam.m
@@ -138,9 +138,9 @@ otherParams              = struct();
 otherParams.problemName  = 'plastic_2dframe';
 % otherParams.plots_format = 'vtk' ;
 
-[modelCurrSol, modelProperties, BCsData] = ONSAS_init(materials, elements, boundaryConds, initialConds, mesh, analysisSettings, otherParams);
+[modelCurrSol, modelProperties, BCsData] = initONSAS(materials, elements, boundaryConds, initialConds, mesh, analysisSettings, otherParams);
 
-[matUs, loadFactorsMat, ~] = ONSAS_solve(modelCurrSol, modelProperties, BCsData);
+[matUs, loadFactorsMat, ~] = solveONSAS(modelCurrSol, modelProperties, BCsData);
 
 girosUltimoNodo_10 = matUs((num_elem + 1) * 6, :);
 descensosUltimoNodo_10 = matUs((num_elem + 1) * 6 - 3, :);
