+++ src/core/assembler.m
@@ -214,7 +214,7 @@ function [fsCell, stressMat, tangMatsCell, localInternalForces, strain_vec, acum
 
         params_plastic_2Dframe(elem, :) = previousStateCell(elem, :);
 
-        [fs, ks, fintLocCoord, aux] = frame2D_plastic_internal_force(elemNodesxyzRefCoords, ...
+        [fs, ks, fintLocCoord, aux] = frame2DPlasticInternalForce(elemNodesxyzRefCoords, ...
                                                                      elemCrossSecParams, ...
                                                                      modelParams, ...
                                                                      elemDisps, params_plastic_2Dframe(elem, :), elem);
