+++ src/core/assembler.m
@@ -215,9 +215,9 @@ function [fsCell, stressMat, tangMatsCell, localInternalForces, strain_vec, acum
         params_plastic_2Dframe(elem, :) = previousStateCell(elem, :);
 
         [fs, ks, fintLocCoord, aux] = frame2DPlasticInternalForce(elemNodesxyzRefCoords, ...
-                                                                     elemCrossSecParams, ...
-                                                                     modelParams, ...
-                                                                     elemDisps, params_plastic_2Dframe(elem, :), elem);
+                                                                  elemCrossSecParams, ...
+                                                                  modelParams, ...
+                                                                  elemDisps, params_plastic_2Dframe(elem, :), elem);
 
         Nx = 0;
         My = 0;
