+++ src/interface/initONSAS.m
@@ -104,6 +104,8 @@ function [modelCurrSol, modelProperties, BCsData] = initONSAS(materials, element
   previousStateCell(:, 2) = {zeros(1, 3)};
   previousStateCell(:, 3) = {0};
 
+  previousPlasticFrameState = zeros(size(Conec, 1), 11);
+
   % ==========================================================================================
   % ==========================================================================================
   % Global rotation matrices for triangle elements
@@ -114,17 +116,17 @@ function [modelCurrSol, modelProperties, BCsData] = initONSAS(materials, element
   % ==========================================================================================
 
   % comput internal forces and stresses
-  [~, Stress, ~, localInternalForces, strain_vec, acum_plas_strain_vec] = assembler(modelProperties.Conec, modelProperties.elements, modelProperties.Nodes, modelProperties.materials, BCsData.KS, U, Udot, Udotdot, modelProperties.analysisSettings, [0 1 0 1], modelProperties.nodalDispDamping, currTime, previousStateCell, rotMatCell);
+  [~, Stress, ~, localInternalForces, strain_vec, acum_plas_strain_vec] = assembler(modelProperties.Conec, modelProperties.elements, modelProperties.Nodes, modelProperties.materials, BCsData.KS, U, Udot, Udotdot, modelProperties.analysisSettings, [0 1 0 1], modelProperties.nodalDispDamping, currTime, previousStateCell, rotMatCell, previousPlasticFrameState);
 
   [FextG, currLoadFactorsVals]  = computeFext(modelProperties, BCsData, 0, length(U), [], {U, Udot, Udotdot});
 
   nextTime = currTime + analysisSettings.deltaT;
 
   % md call assembler
-  [systemDeltauMatrix, systemDeltauRHS, ~, ~, ~, ~, modelProperties.exportFirstMatrices] = systemAssembler(modelProperties, BCsData, U, Udot, Udotdot, U, Udot, Udotdot, nextTime, [], previousStateCell, rotMatCell);
+  [systemDeltauMatrix, systemDeltauRHS, ~, ~, ~, ~, modelProperties.exportFirstMatrices] = systemAssembler(modelProperties, BCsData, U, Udot, Udotdot, U, Udot, Udotdot, nextTime, [], previousStateCell, rotMatCell, previousPlasticFrameState);
 
   modelCurrSol = constructModelSol(timeIndex, currTime, U, Udot, Udotdot, Stress, convDeltau, ...
-                                   currLoadFactorsVals, systemDeltauMatrix, systemDeltauRHS, timeStepStopCrit, timeStepIters, localInternalForces, previousStateCell, rotMatCell);
+                                   currLoadFactorsVals, systemDeltauMatrix, systemDeltauRHS, timeStepStopCrit, timeStepIters, localInternalForces, previousStateCell, rotMatCell, previousPlasticFrameState);
   % =================================================================
 
   % md prints headers for solver output file
