+++ src/constructors/constructModelSol.m
@@ -19,7 +19,7 @@ function modelCurrSol = constructModelSol(timeIndex, currTime, U, Udot, ...
                                           Udotdot, Stress, convDeltau, ...
                                           currLoadFactorsVals, systemDeltauMatrix, ...
                                           systemDeltauRHS, timeStepStopCrit, timeStepIters, ...
-                                          localInternalForces, previousStateCell, rotMatCell)
+                                          localInternalForces, previousStateCell, rotMatCell, previousPlasticFrameState)
 
   modelCurrSol = struct ('timeIndex', timeIndex, ...
                          'currTime', currTime, ...
@@ -37,3 +37,4 @@ function modelCurrSol = constructModelSol(timeIndex, currTime, U, Udot, ...
   modelCurrSol.localInternalForces = localInternalForces;
   modelCurrSol.previousStateCell = previousStateCell;
   modelCurrSol.rotMatCell = rotMatCell;
+  modelCurrSol.previousPlasticFrameState = previousPlasticFrameState;

[…]
