+++ src/core/timeStepIteration.m
@@ -167,7 +167,7 @@ function modelNextSol = timeStepIteration(modelCurrSol, modelProperties, BCsData
   modelNextSol = constructModelSol(timeIndex, currTime, U, Udot, ...
                                    Udotdot, Stress, convDeltau, ...
                                    nextLoadFactorsVals, systemDeltauMatrix, ...
-                                   systemDeltauRHS, timeStepStopCrit, timeStepIters, matFint, previousStateCell, rotMatCell, previousPlasticFrameState);
+                                   systemDeltauRHS, timeStepStopCrit, timeStepIters, matFint, previousStateCell, rotMatCell, frameStateCellnp1);
 
   % ==============================================================================
   % ==============================================================================
