+++ src/core/assembler.m
@@ -213,7 +213,6 @@ function [fsCell, stressMat, tangMatsCell, localInternalForces, strain_vec, acum
 
       elseif strcmp(modelName, 'plastic-2Dframe')
 
-
         [fs, ks, fintLocCoord, aux] = frame2DPlasticInternalForce(elemNodesxyzRefCoords, ...
                                                                   elemCrossSecParams, ...
                                                                   modelParams, ...
