+++ examples/plasticDarvallMendisFrame/plasticDarvallMendisFrame.m
@@ -19,8 +19,7 @@
 close all;
 clear all;
 
-addpath(genpath([pwd '/../src/'])); % add ONSAS directory to path
-addpath(genpath([pwd '/../../ONSAS/src'])); % add ONSAS directory to path
+addpath(genpath([pwd '/../../src/']));
 
 global TZERO
 
@@ -133,7 +132,7 @@ initialConds = struct();
 analysisSettings                    = {};
 analysisSettings.methodName         = 'arcLength';
 analysisSettings.deltaT             = 1;
-analysisSettings.incremArcLen       = [1e-5 * ones(1, 2)];
+analysisSettings.incremArcLen       = [1e-5 * ones(1, 2100)];
 analysisSettings.finalTime          = length(analysisSettings.incremArcLen);
 analysisSettings.iniDeltaLamb       = 1;
 analysisSettings.posVariableLoadBC  = 2;
@@ -182,419 +181,10 @@ fprintf(' First hinge\n Displacement u = %g\n Load Factor lambdaF = %g\n', matUs
 fprintf(' Second hinge\n Displacement u = %g\n Load Factor lambdaF = %g\n', matUs((4) * 6 + 3, TZERO(7)) * 100, loadFactorsMat(TZERO(7), 2));
 fprintf(' Third hinge\n Displacement u = %g\n Load Factor lambdaF = %g\n', matUs((4) * 6 + 3, TZERO(2)) * 100, loadFactorsMat(TZERO(2), 2));
 
-% /\  /\  /\  /\  /\  /\  /\  /\  /\  /\  /\  /\  /\  /\  /\  /\  /\  /\
+numSecHingeDisp = matUs((4) * 6 + 3, TZERO(7)) * 100;
+numSecHingeLoad = loadFactorsMat(TZERO(7), 2);
 
-% geometry
-L  = 3.048;            % m
-L1 = 0.55 * L;
+refDisp = -1.14;
+refForce = 427;
 
-Inertia = 0.001;       % m^4
-
-A = 0.103;            % m^2
-
-% material
-E = 20.68e6;          % kN/m^2 [kPa]                                                                                                                                                % kN/m^2 [kPa]
-
-nu = 0.3;     % Poisson's ratio
-
-a = -0.04;
-% a = -0.06 ;
-% a = -0.0718 ;
-
-kh1 = 30000;               % kN.m^2
-kh2 = 3000;
-
-Ks = a * E * Inertia / L1 * 10;    % kN.m
-
-Mc_c = 158.18;             % kN.m
-My_c = 158.18;
-Mu_c = 158.18;
-
-Mc_b = 169.48;             % kN.m
-My_b = 169.48;
-Mu_b = 169.48;
-
-% /\  /\  /\  /\  /\  /\  /\  /\  /\  /\  /\  /\  /\  /\  /\  /\  /\  /\
-
-materials = struct();
-materials(1).modelName = 'plastic-2Dframe';
-materials(1).modelParams = [E Mc_c My_c Mu_c kh1 kh2 Ks nu];
-materials(2).modelName = 'plastic-2Dframe';
-materials(2).modelParams = [E Mc_b My_b Mu_b kh1 kh2 Ks nu];
-% elements
-elements = struct();
-elements(1).elemType  = 'node';
-elements(2).elemType  = 'frame';
-elements(2).elemCrossSecParams = {'generic'; [A 1 Inertia Inertia] };
-% Boundary Conditions
-% Supports
-boundaryConds = struct();
-boundaryConds(1).imposDispDofs = [1 2 3 4 5 6];
[…]
