+++ test/runTestProblemsLocal.m
@@ -34,6 +34,7 @@ keyfiles = { ...
             ; 'frameLinearAnalysis/frameLinearAnalysis.m'               ...
             ; 'linearAerodynamics/linearAerodynamics.m'                 ...
             ; 'nonLinearPendulum/nonLinearPendulum.m'                   ...
+            ; 'plasticDarvallMendisFrame/plasticDarvallMendisFrame.m'   ...
             ; 'platePatchTest/platePatchTest.m'                         ...
             ; 'ringPlaneStrain/ringPlaneStrain.m'                       ...
             ; 'simplePropeller/simplePropeller.m'                       ...
