+++ examples/plasticCantileverBeam/softHinge1DOFSemiAnalyticSol.m
@@ -1,145 +1,0 @@
-% Copyright 2025, ONSAS Authors (see documentation)
-%
-% This file is part of ONSAS.
-%
-% ONSAS is free software: you can redistribute it and/or modify
-% it under the terms of the GNU General Public License as published by
-% the Free Software Foundation, either version 3 of the License, or
-% (at your option) any later version.
-%
-% ONSAS is distributed in the hope that it will be useful,
-% but WITHOUT ANY WARRANTY; without even the implied warranty of
-% MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-% GNU General Public License for more details.
-%
-% You should have received a copy of the GNU General Public License
-% along with ONSAS.  If not, see <https://www.gnu.org/licenses/>.
-
-% algorithm for the update of the internal variables for elastoplasticity with hardening
-% and for the computation of the moment in the bulk Mn1
-
-% displacements in time n + 1, dpn1 v1, v2, theta1, theta2, alpha, xd
-% plastic curvature in time n / kappa_plas_n
-
-function [kappa_plas_n1, xin11val, xin21val, alfan1, Mn1] = softHinge1DOFSemiAnalyticSol(v1, v2, theta1, theta2, xd, alfan, xin1, xin2, kappa_plas_n, Mc, My, Mu, kh1, kh2, Ks, E, Iy, l)
-
-  % integration points
-  x = [0 l / 2 l];
-  wp = [1 / 3 4 / 3 1 / 3] * l * 0.5;
-
-  npi = length(x);
-
-  tM = 0;
-
-  Bv1 = -6 / l^2 * (1 - 2 * x / l);
-  Bv2 =  6 / l^2 * (1 - 2 * x / l);
-
-  Bt1 = -2 / l * (2 - 3 * x / l);
-  Bt2 = -2 / l * (1 - 3 * x / l);
-
-  G_bar = -(1 + 3 * (1 - 2 * xd / l) * (1 - 2 * x / l)) / l;
-
-  kappa_plas_test = kappa_plas_n;
-
-  % smooth curvature
-  kappa_bar = Bv1 * v1 + Bv2 * v2 + Bt1 * theta1 + Bt2 * theta2 + G_bar * alfan;
-
-  Mn1 = E * Iy * (kappa_bar - kappa_plas_test);
-
-  % /\ /\ /\ /\ /\ /\ /\ /\ /\ /\ /\ /\ /\ /\ /\ /\ /\ /\ /\ /\ /\ /\ /\ /\
-
-  % hardening
-
-  % hardening function
-  if xin1 <= (My - Mc) / kh1
-    q = -kh1 * xin1;
-  else
-    q = -(My - Mc) * (1 - kh2 / kh1) - kh2 * xin1;
-  end
-
-  % yield function test
-  phi_test = abs(Mn1) - (Mc - q);
-
-  if phi_test <= 0
-
-    kappa_plas_n1 = kappa_plas_n;
-    xin11val = xin1;
-
-  else
-
-    % calculation of gamma_n1
-    if xin1 + phi_test / (kh1 + E * Iy) <= (My - Mc) / kh1
-
-      gamma_n1 = phi_test / (kh1 + E * Iy);
-
-    else
-
-      gamma_n1 = phi_test / (kh2 + E * Iy);
-
-    end
-
-    kappa_plas_n1 = kappa_plas_n + gamma_n1 .* sign(Mn1);
-    xin11val = xin1 + gamma_n1;
-
-  end
-
[…]
