+++ examples/plasticCantileverBeam/softHinge1DOFNumericSol.m
@@ -1,145 +1,0 @@
-% Copyright 2025, ONSAS Authors (see documentation)
-%
-% This file is part of ONSAS.
-%
-% ONSAS is free software: you can redistribute it and/or modify
-% it under the terms of the GNU General Public License as published by
-% the Free Software Foundation, either version 3 of the License, or
-% (at your option) any later version.
-%
-% ONSAS is distributed in the hope that it will be useful,
-% but WITHOUT ANY WARRANTY; without even the implied warranty of
-% MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-% GNU General Public License for more details.
-%
-% You should have received a copy of the GNU General Public License
-% along with ONSAS.  If not, see <https://www.gnu.org/licenses/>.
-
-% numerical example
-% cantilever beam loaded with a vertical force at the free end
-
-function [Mn, Fn, matdes] = softHinge1DOFNumericSol(l, A, E, Inertia, Mc, My, Mu, kh1, kh2, Ks)
-
-  freedofs = [2 4 6]; % u2 v2 theta2
-
-  % at the beginning..., there was no softening hinge
-  soft_hinge_boolean = false;
-
-  % Gauss-Lobatto Quadrature with 3 integration points [a (a+b)/2 b]
-  npi = 3;
-  xpi = [0 l / 2 l];
-  wpi = [1 / 3 4 / 3 1 / 3] * l * 0.5;
-
-  nu   = 0.3;
-  tol1 = 1e-8;
-  tol2 = 1e-8;
-  tolk = 15;
-
-  % initial values
-  dn   = [0 0 0 0 0 0]';
-  Fint = [0 0 0 0 0 0]';
-  tM   = 0;
-
-  kpn  = zeros(npi, 1);
-  xin1 = zeros(npi, 1);
-  xin2 = 0;
-
-  kpn1  = zeros(npi, 1);
-  xin11 = zeros(npi, 1);
-  xin21 = 0;
-
-  khat1 = zeros(npi, 1);
-
-  M1 = zeros(npi, 1);
-  Fn = zeros(npi, 1);
-
-  alfan = 0;
-
-  xd = 0;
-  xdi = 1;
-
-  Final_force = 205; % value of the final force
-
-  load_case = [0 0 0 1 0 0]'; % load applied in vertical direction (Y)
-  load_factors = 0:Final_force;
-
-  % --- element params ---
-  elemParams = [l A Inertia];
-
-  % --- elastoplastic params ---
-  elastoplasticParams = [E Mc My Mu kh1 kh2 Ks];
-
-  matdes = zeros (6, Final_force + 1);
-
-  matdes(:, 1) = dn;
-
-  gxin = zeros(Final_force, 1);
-  gxin2 = zeros(Final_force, 1);
-  gkpn = zeros(Final_force, 1);
-
-  Mn = zeros(Final_force, 1);
-  TM = zeros(Final_force, 1);
-
-  Alf = zeros(Final_force, 1);
-
-  for ind = 2:length(load_factors)
-
-    curr_load_factor = load_factors(ind);
-
-    Fext = load_case * curr_load_factor;
-
-    dnk = matdes(:, ind - 1);
-
-    % iteration vars
-    converged_boolean = false;
-    k = 0; % set iterations zero
-
-    gxin(ind - 1, 1)   = xin1(1);
-    gxin2(ind - 1, 1)  = xin2;
-    gkpn(ind - 1, 1)   = kpn(1);
-    Mn(ind - 1, 1)     = M1(1);
-    TM(ind - 1, 1)     = tM;
-    Fn(ind - 1, 1)     = Fint(4);
[…]
