+++ examples/plasticCantileverBeam/plasticCantileverBeam.m
@@ -1,575 +1,0 @@
-% Copyright 2025, ONSAS Authors (see documentation)
-%
-% This file is part of ONSAS.
-%
-% ONSAS is free software: you can redistribute it and/or modify
-% it under the terms of the GNU General Public License as published by
-% the Free Software Foundation, either version 3 of the License, or
-% (at your option) any later version.
-%
-% ONSAS is distributed in the hope that it will be useful,
-% but WITHOUT ANY WARRANTY; without even the implied warranty of
-% MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-% GNU General Public License for more details.
-%
-% You should have received a copy of the GNU General Public License
-% along with ONSAS.  If not, see <https://www.gnu.org/licenses/>.
-
-% =========================================================================
-
-% Euler-Bernoulli element with embedded discontinuity
-
-% Numerical modeling of softening hinges in thin Euler–Bernoulli beams
-% Francisco Armero, David Ehrlich / University of California, Berkeley
-
-% Embedded discontinuity finite element formulation
-% For failure analysis of planar reinforced concrete beams and frames
-% Miha Jukić, Boštjan Brank / University of Ljubljana
-% Adnan Ibrahimbegović / Ecole normale supérieure de Cachan
-
-% =========================================================================
-
-% numerical example
-% cantilever beam loaded with a vertical force at the free end
-
-% =========================================================================
-
-close all;
-clear;
-addpath(genpath([pwd '/../../src']));
-
-% assumed XY plane
-
-% -------------------------------------------
-% scalar parameters
-% material
-EI  = 77650;       % KN.m^2
-kh1 = 29400;       % KN.m^2
-kh2 = 273;
-Ks  = -18000;      % KN.m
-
-nu = 0.3;          % Poisson's ratio
-
-% geometry
-l  = 2.5;              % m
-ty = 0.3;              % width cross section
-tz = 0.4;              % height cross section
-Inertia = tz * ty^3 / 12;  % m^4
-
-E = EI / Inertia;        % KN/m^2 [KPa]
-
-A  = ty * tz;            % m^2
-Mc = 37.9;             % KN.m
-My = 268;
-Mu = 374;
-
-% /\   /\   /\   /\   /\   /\   /\   /\   /\   /\   /\   /\   /\   /\   /\
-% ONSAS (NUMBER OF ELEMENTS 10)
-
-% at the beginning..., there was no softening hinge
-% soft_hinge_boolean = false ;
-
-% number of finite elements
-num_elem = 10;
-
-materials             = struct();
-materials.modelName   = 'plastic-2Dframe';
-materials.modelParams = [E Mc My Mu kh1 kh2 Ks nu];
-
-elements             = struct();
-elements(1).elemType = 'node';
-
-elements(2).elemType = 'frame';
-elements(2).elemCrossSecParams = {'generic'; [A 1 Inertia Inertia] };
-
-boundaryConds                  = {};
[…]
