+++ previews/PR375/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"corotationalFrameElement/#Co-rotational-frame-system","page":"Corotational frame","title":"Co-rotational frame system","text":"","category":"section"},{"location":"corotationalFrameElement/","page":"Corotational frame","title":"Corotational frame","text":"<img src=\"https://raw.githubusercontent.com/ONSAS/ONSAS_docs/master/docs/src/diagramCorotationalFrameSystem.svg\" alt=\"corotational diagram\" width=\"800\"/>","category":"page"},{"location":"tutorials/CantileverBeam/cantileverBeam/","page":"-","title":"-","text":"%## Uniform curvature cantilever beam %#–- %# %#In this tutorial, the Uniform curvature cantilever example and its resolution using ONSAS are described. The aim of this example is to validate the static co-rotational 3D beam implementation by comparing the results provided by ONSAS with the analytical solution.  The Octave script of this example is available at this url. %# %#The problem consists in a beam, with one free end (right) submitted to a nodal moment M, and the other end (left) constrained (welded), as it is shown in the figure. %# %#@raw html %#<img src=\"https://raw.githubusercontent.com/ONSAS/ONSAS_docs/master/docs/src/tutorials/CantileverBeam/cantileverBeam.svg\" alt=\"structure diagram\" width=\"500\"/> %# %# %#Before defining the structs, the workspace is cleaned, the ONSAS directory is added to the path and scalar geometry and material parameters are defined. close all, clear all ; % add path addpath( [ pwd '/../../src'] ); % material scalar parameters E = 200e9 ;  nu = 0.3 ; % geometrical scalar parameters l = 10 ; ty = .1 ;  tz = .1 ; % the number of elements of the mesh numElements = 10 ; %# %### MEBI parameters %#––––––––– %# %#The modelling of the structure begins with the definition of the Material-Element-BoundaryConditions-InitialConditions (MEBI) parameters. %#### materials %# Since the example contains only one rod the fields of the materials struct will have only one entry. Although, it is considered constitutive behavior according to the SaintVenantKirchhoff law: materials.hyperElasModel  = { '1DrotEngStrain'} ; materials.hyperElasParams = { [ E nu ] } ; %# The density is not defined, therefore it is considered as zero (default), then no inertial effects are considered (static analysis). %# %#### elements %# %#Two different types of elements are considered, node and beam. The nodes will be assigned in the first entry (index 1) and the beam at the index 2. The elemType field is then: elements.elemType = { 'node','frame' } ; %# for the geometries, the node has not geometry to assign (empty array), and the truss elements will be set as a rectangular-cross section with t_y and t_z cross-section dimensions in y and z directions, then the elemTypeGeometry field is: elements.elemTypeGeometry = { [], [2 ty tz ] }; elements.elemTypeParams = { [], 1 }; %### boundaryConds %# %# The elements are submitted to two different BC settings. The first BC corresponds to a welded condition (all 6 dofs set to zero), and the second corresponds to an incremental nodal moment, where the target load produces a circular form of the deformed beam.  %# The scalar values of inertia I_z is computed. Iy = tytz^3/12 ; boundaryConds.loadsCoordSys = { []        ; 'global'   } ; boundaryConds.loadsTimeFact = { []        ; @(t) EIy2pi/l t } ; boundary
[…]
