+++ docs/src/index.md
@@ -9,7 +9,7 @@
 
 [ONSAS](https://github.com/ONSAS/ONSAS) is an Open Nonlinear Structural Analysis Solver for GNU-Octave/Matlab. It consists in a set of implementations of numerical methods for static/dynamic and linear/non-linear analysis of structures. The first version was developed for educational purposes and published in a Structural Analysis [handbook](https://www.colibri.udelar.edu.uy/jspui/bitstream/20.500.12008/22106/1/Bazzano_P%c3%a9rezZerpa_Introducci%c3%b3n_al_An%c3%a1lisis_No_Lineal_de_Estructuras_2017.pdf) for a graduate course taught at the [School of Engineering of Universidad de la República](https://www.fing.edu.uy/). The current version allows to perform a diverse set of simulations and it has been used in specific research applications.
 
-## Publications using ONSAS
+## Some Publications using ONSAS
 
 Journal articles using ONSAS:
 
@@ -21,6 +21,10 @@ Theses using ONSAS:
 
  * 2021, M. Vanzulli, *Implementación de una formulación corrotacional en dinámica no lineal y aplicación al modelado de líneas de transmisión eléctrica* [url](https://www.colibri.udelar.edu.uy/jspui/handle/20.500.12008/28388)
  * 2021, A. Teliz, *Optimización de torres de alta tensión y su análisis frente a vientos de alta intensidad* [url](https://hdl.handle.net/20.500.12008/35985)
+
+Conference full-articles using ONSAS:
+
+ * 2025, S. Merlino, J.M. Pérez Zerpa, *Implementación en Código Abierto de un Método para el Análisis Post-Colapso de Pórticos*, MECOM AMCA [proceeding](https://amcaonline.org.ar/ocs/index.php/mecom2025/mecom2025/paper/view/8246)
 
 
 ## Some academic examples
@@ -80,6 +84,8 @@ The authorship of each version is (or tends to be) based on the criteria defined
 * [**Mauricio Vanzulli**](https://github.com/mvanzulli) (**2**) co-developed the Newmark's method functions and scripts, developed input files for the dynamic analysis examples. Developed the nonlinear dynamic co-rotational frame element function, its validation and integration with the VIV function. 
 
 * [**Joaquín Viera**](https://exportcvuy.anii.org.uy/cv/?b6b1cd2fe90a9c29279eedb0d3cc4c4d) (**1**), leaded the development of the Linear Analysis module and input files, collaborated in the design and development of the input reading and output generation modules, leaded the development of a GUI.
+
+* [**Sergio Merlino**](**1**), leaded the development of the formulation for plastic analysis of frames considering hardening and softening hinges.
 
 * [**Alexandre Villié**](https://www.linkedin.com/in/alexandre-villi%C3%A9-343870187/) (**3**) developed the current Vortex-Induced-Vibrations Wake Oscillator model. Also contributed in the validation of this function in the integration with the co-rotational frame element.
 
