+++ CITATION.cff
@@ -13,6 +13,8 @@ authors:
 - family-names: "Schaedler de Almeida"
   given-names: "F."
   orcid: "https://orcid.org/0000-0003-1535-5405"
+- family-names: "Merlino Chiozza"
+  given-names: "S. A."
 - family-names: "Villié"
   given-names: "A."
   orcid: "https://orcid.org/0000-0001-5431-2892" 
