+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Name: ONSAS
-Version: 0.3.3
-Date: 2025-04-27
-Author: Pérez Zerpa, J. M. and Vanzulli, M. and Viera, J. and Villie, A. and Forets, M. and Schaedler de Almeida, F. and Battini, J.-M. and Bazzano, J. B. 
+Version: 0.5.0
+Date: 2025-11-22
+Author: Pérez Zerpa, J. M. and Vanzulli, M. and Viera, J. and Merlino S. and Villie, A. and Forets, M. and Schaedler de Almeida, F. and Battini, J.-M. and Bazzano, J. B. 
 Maintainer: Pérez Zerpa, J. M. (jorgepz@fing.edu.uy)
 Title: ONSAS: an Open Nonlinear Structural Analysis Solver
 Description: ONSAS is a GNU-Octave code for static/dynamic and linear/non-linear analysis of solids, beams, truss or plane structures.
