+++ CITATION.cff
@@ -27,5 +27,5 @@ authors:
 - family-names: "Bazzano García"
   given-names: "J. B."
 title: "ONSAS: an Open Nonlinear Structural Analysis Solver for GNU-Octave/Matlab"
-version: 0.3.3
+version: 0.4.1
 url: "https://github.com/ONSAS/ONSAS"
