+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Name: ONSAS
-Version: 0.5.0
+Version: 0.4.1
 Date: 2025-11-22
 Author: Pérez Zerpa, J. M. and Vanzulli, M. and Viera, J. and Merlino S. and Villie, A. and Forets, M. and Schaedler de Almeida, F. and Battini, J.-M. and Bazzano, J. B. 
 Maintainer: Pérez Zerpa, J. M. (jorgepz@fing.edu.uy)
