+++ src/interface/initONSAS.m
@@ -27,7 +27,7 @@ function [modelCurrSol, modelProperties, BCsData] = initONSAS(materials, element
   [materials, elements, boundaryConds, analysisSettings, otherParams] = setDefaults(materials, elements, boundaryConds, analysisSettings, otherParams);
 
   % md sets the current version and welcomes user
-  ONSASversion = '0.5.0';
+  ONSASversion = '0.4.1';
   welcomeMessage(ONSASversion, otherParams);
 
   % creates outputdir in current location
