+++ test/runTestProblemsMoxunitDisplacements.m
@@ -112,4 +112,8 @@ function test20
   assertEqual(verifBoolean, true);
 
 function test21
+  plasticDarvallMendisFrame;
+  assertEqual(verifBoolean, true);
+
+function test22
   assertEqual(gaussIntegrationTest, true);
