+++ docs/src/bringONSASmFilesToONSASdocs.m
@@ -16,12 +16,7 @@
 % along with ONSAS.  If not, see <https://www.gnu.org/licenses/>.
 %
 
-% md function used to convert .m files in the ONSAS repo to .md files in this repo
-% md this function is executed from the makeAndPreview.sh file or can also be
-% md executed from the bash terminal: `octave --eval "bringONSASmFilesToONSAS_docs('$ONSAS_PATH')"`
-% md where \$ONSAS_PATH is the environment variable with the directory of ONSAS
-
-function bringONSASmFilesToONSASdocs
+function bringONSASmFilesToONSASdocs()
   disp('running bringONSASmFilesToONSASdocs script...');
   dirONSASdocs = './examples/';
   dirONSASm     = '../../examples/';
