+++ src/core/convergenceTest.m
@@ -17,7 +17,6 @@
 %
 function [booleanConverged, stopCritPar, deltaErrLoad, normFext] = convergenceTest( ...
                                                                                    analysisSettings, redFext, redDeltaU, redUk, dispIter, systemDeltauRHS)
-
   stopTolDeltau = analysisSettings.stopTolDeltau;
   stopTolForces = analysisSettings.stopTolForces;
   stopTolIts    = analysisSettings.stopTolIts;

[…]
