### 2025-11-22 7bdd0c8
Update paper.bib
