+++ paper/paper.bib
@@ -1,5 +1,18 @@
 
 
+@article{jukic,
+title = {Embedded discontinuity finite element formulation for failure analysis of planar reinforced concrete beams and frames},
+journal = {Engineering Structures},
+volume = {50},
+pages = {115-125},
+year = {2013},
+note = {Engineering Structures: Modelling and Computations (special issue IASS-IACM 2012)},
+issn = {0141-0296},
+doi = {https://doi.org/10.1016/j.engstruct.2012.07.028},
+url = {https://www.sciencedirect.com/science/article/pii/S0141029612004002},
+author = {Miha Jukić and Boštjan Brank and Adnan Ibrahimbegović}
+}
+
 @book{Zienkiewicz1972,
 author = {Zienkiewicz, Olgierd C.},
 booktitle = {Introductory Lectures on the Finite Element Method},
