### 2025-11-22 #8473 @jorgepz
Thanks @AnjaliSandip for accepting the editor invitation and thanks @kyleniemeyer for inviting Anjali!

Regarding potential reviewers, considering the language and the topic areas I consider either Kevin-Mattheus-Moerman, Damir Akchurin or Onur Deniz Akan could contribute to the reviewing process.

Finally I would like to mention that luckily ONSAS has recieved valuable contributions in the last months, and the manuscript was slightly updated to reflect this. I will regenerate it.
