### 2025-11-22 bd2ec2b
Update paper.bib
