+++ paper/paper.bib
@@ -8,7 +8,7 @@ pages = {115-125},
 year = {2013},
 note = {Engineering Structures: Modelling and Computations (special issue IASS-IACM 2012)},
 issn = {0141-0296},
-doi = {https://doi.org/10.1016/j.engstruct.2012.07.028},
+doi = {10.1016/j.engstruct.2012.07.028},
 url = {https://www.sciencedirect.com/science/article/pii/S0141029612004002},
 author = {Miha Jukić and Boštjan Brank and Adnan Ibrahimbegović}
 }
