+++ paper/paper.bib
@@ -1,262 +1,0 @@
-
-
-@article{jukic,
-title = {Embedded discontinuity finite element formulation for failure analysis of planar reinforced concrete beams and frames},
-journal = {Engineering Structures},
-volume = {50},
-pages = {115-125},
-year = {2013},
-note = {Engineering Structures: Modelling and Computations (special issue IASS-IACM 2012)},
-issn = {0141-0296},
-doi = {https://doi.org/10.1016/j.engstruct.2012.07.028},
-url = {https://www.sciencedirect.com/science/article/pii/S0141029612004002},
-author = {Miha Jukić and Boštjan Brank and Adnan Ibrahimbegović}
-}
-
-@book{Zienkiewicz1972,
-author = {Zienkiewicz, Olgierd C.},
-booktitle = {Introductory Lectures on the Finite Element Method},
-doi = {10.1007/978-3-7091-2973-9},
-title = {Introductory Lectures on the Finite Element Method},
-year = {1972}
-}
-
-@article{Battini2002,
-author = {Battini, Jean Marc and Pacoste, Costin},
-doi = {10.1016/S0045-7825(01)00352-8},
-issn = {00457825},
-journal = {Computer Methods in Applied Mechanics and Engineering},
-number = {17-18},
-pages = {1755--1789},
-title = {Co-rotational beam elements with warping effects in instability problems},
-volume = {191},
-year = {2002}
-}
-
-@article{yoshi,
-title = {Elastic buckling phenomenon applicable to deployable rings},
-journal = {International Journal of Solids and Structures},
-volume = {29},
-number = {7},
-pages = {893-909},
-year = {1992},
-issn = {0020-7683},
-doi = {10.1016/0020-7683(92)90024-N},
-url = {https://www.sciencedirect.com/science/article/pii/002076839290024N},
-author = {Goto Yoshiaki and Watanabe Yasuhito and Kasugai Toshihiro and Obata Makoto}
-}
-
-@book{Bathe1982,
-author = {Bathe, Klaus-J{\"{u}}rgen},
-isbn = {978-0133173055},
-pages = {736},
-publisher = {Prentice-Hall},
-title = {{Finite Element Procedures in Engineering Analysis}},
-year = {1982}
-}
-
-@book{Crisfield1997,
-annote = {0020},
-author = {Crisfield, Mike A.},
-isbn = {978-0471956495},
-pages = {508},
-publisher = {Wiley},
-title = {{Non-Linear Finite Element Analysis Solids and Structure, Volume 2, Advanced Topics}},
-year = {1997}
-}
-
-@article{Battini2002,
-	author = {Battini, Jean Marc and Pacoste, Costin},
-	doi = {10.1016/S0045-7825(01)00352-8},
-	issn = {00457825},
-	journal = {Computer Methods in Applied Mechanics and Engineering},
-	keywords = {Finite element and matrix methods,Nonlinear mechanics,Numerical solution procedures,Stability in structural mechanics,Structural mechanics},
-	number = {17-18},
-	pages = {1755--1789},
-	title = {{Co-rotational beam elements with warping effects in instability problems}},
-	volume = {191},
-	year = {2002}
-}
-
-@article{Le2014,
-	author = {Le, Thanh Nam and Battini, Jean Marc and Hjiaj, Mohammed},
-	journal = {Computer Methods in Applied Mechanics and Engineering},
-	keywords = {3D beam elements,Corotational method,Finite rotations,Nonlinear dynamics},
-	pages = {538--565},
-	doi = {10.1016/j.cma.2013.11.007},
-	title = {{A consistent 3D corotational beam element for nonlinear dynamic analysis of flexible structures}},
-	volume = {269},
-	year = {2014}
-}
-
[…]
