### 2025-12-10 #9604 @editorialbot
[REVIEW]: ONSAS: an Open Nonlinear Structural Analysis Solver for GNU-Octave/MATLAB
**Submitting author:** @jorgepz (<a href="http://orcid.org/0000-0002-8778-624X">Jorge Martín Pérez Zerpa</a>)
**Repository:** https://github.com/ONSAS/ONSAS
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.3.3
