### 2026-01-29 #9604 @akchurinda
Dear @jorgepz,

This piece of software is a great contribution to the Structural Engineering field.

Before I proceed further with my review, I would like you to address some of my suggestions:

- Although not required, I think the paper would greatly benefit from more in-depth discussion on the State of Field - comparison of the library of material, section, and elements formulations contained in the ONSAS against the existing packages, such as OpenSees(Py) which you reference. Mentioning other existing packages (there are a lot more than two mentioned) for structural analysis would definitely be useful too.
- One thing that has always bothered me about the existing packages for structural analysis, from the research point of view, is the lack of description of element formulations (even brief) provided in the packages. Even a single frame element can be implemented many different ways based on the formulation (Euler-Bernoulli, Timoshenko-Ehrenfest, fiber-based, etc.). OpenSees(Py) is an example of such package, which shows how to use the package but barely provides information about the underlying element formulations forcing the users to infer them from the codebase. I think your package and its community would greatly benefit from a separate section in the documentation that, if does not provide full blown formulations which understandably can be very long, at least does a brief summary for each element and provides useful references. It is a real opportunity for the package to stand out from the rest.
- The package would also benefit from a dedicated `CONTRIBUTING.md` file to provide guidance to the users on how to properly contribute to the package.
- Some minor editorial comments:
  - On line 76, I recommend using "Figure 1 a)" rather than "Figure 1 subfigure a)".
  - On several occasions, in-text citations should be written as "... in Jukic et al. (2013)" rather than "... in (Jukic et al., 2013)". I believe this is easily achieved by dropping square brackets from the citation in the `.md` file: `[@jukic]` to `@jukic`.
  - Please fix missing DOI for items detected by the @editorialbot.

Please let me know when these suggestions are addressed and I will continue my review.
