### 2026-02-10 #9604 @odakan
Dear @AnjaliSandip ,

**Please find a preliminary review of the ONSAS framework, below:**

_1. Verification of Novelty:_

The authors claim that ONSAS provides an open implementation of the consistent dynamic formulation by Le et al. (2014). This is critical for nonlinear elastic applications such as the wind turbine problem offered as an example, as well as aerospace structures, robotics, and multibody dynamics where inertial effects (e.g.,  deformations due to centripetal forces during high-speed rotations) cannot be neglected.

While I am aware of other open-source corotational applications such as the ASDShellQ4 and forBeamColumn elements in OpenSees, my review of its implementation confirms that they utilize diagonalized lumped-mass approach and typically neglects rotatory inertia. This is sufficient for seismic problems but would be lacking for the wind turbine example.

In contrast, my review of the ONSAS source code confirms that it utilizes a consistent mass formulation integrated via Gauss quadrature. The code explicitly computes the Consistent Mass Matrix, and the Gyroscopic Matrix within the interElementBeamForces function. This supports the authors' claim of a novel implementation of the theory put forward by Le et al. (2014).

_2. Observation on Software Architecture:_

While the co-rotational and dynamic formulations are highly advanced, I noted that the element implementation strategy currently couples the material constitutive logic directly within the element formulation. Specifically, the 2D solid element (elementTriangSolid.m) contains the hard-coded logic for J2 plasticity and the Radial Return Mapping algorithm.

Separation of these objects, similar to the API delegation between Elements and Materials found in OpenSees, would significantly enhance the library's modularity and ease of extension for new material models. However, I recognize that implementing such a separation would require major structural changes. Given that ONSAS is written in MATLAB/Octave language, and focuses on educational clarity and research transparency rather than high performance computing, the current approach is understandable and effective for its intended purpose.

I congratulate the authors (@jorgepz) for their efforts and success.

**Next Steps:**
I am now looking into their broader co-rotational formulation (the main contribution of ONSAS) and the provided validation examples.
