### 2026-02-10 #9604 @jorgepz
Thanks [akchurinda](https://github.com/akchurinda) for your comments, I will let you know as soon as I update the paper, which should be by the end of this week. Thanks also to [odakan](https://github.com/odakan) for your comments so far.
