### 2026-02-24 #9604 @jorgepz
Dear @AnjaliSandip sorry for the delay in addressing the comments. By Wednesday night I should have all the comments addressed.

@akchurinda Thank you for your comments, regarding the contributing.md it was added to the master branch in [this PR](https://github.com/ONSAS/ONSAS/pull/793/changes). And regarding the missing DOI of Bathe's book it seems that editorial bot is suggesting adding [the DOI of a review of the book](https://www.sciencedirect.com/science/article/abs/pii/0141029684900142?via%3Dihub) but not the Finite Element Procedures book. I think the book does not have a DOI but I will check that out.

The other comments should be addressed on Monday. Thanks again.
Jorge
