+++ paper/paper.bib
@@ -260,3 +260,24 @@ keywords = {Reachability analysis, Finite element method, Heat transfer, Structu
   year      = {2005},
   note      = {{ISBN}~978-0123875822},
 }
+
+
+@software{PyElastica,
+  author       = {Arman Tekinalp and
+                  Seung Hyun Kim and
+                  Yashraj Bhosale and
+                  Tejaswin Parthasarathy and
+                  Noel Naughton and
+                  Ali Albazroun and
+                  Rahul Joon and
+                  Songyuan Cui and
+                  Ilia Nasiriziba and
+                  Maximilian Stölzle and
+                  Chia-Hsien (Cathy) Shih and
+                  Mattia Gazzola},
+  title        = {GazzolaLab/PyElastica},
+  year         = 2024,
+  publisher    = {Zenodo},
+  doi          = {10.5281/zenodo.7658871},
+  url          = {https://doi.org/10.5281/zenodo.7658871}
+}
