+++ paper/paper.md
@@ -51,22 +51,26 @@ bibliography: paper.bib
 
 # Summary
 
-The design of structures relies on the computation (or estimation) of the stress and deformation developed by structural elements submitted to external loads. The Finite Element Method (FEM) can be considered the most effective computational tool for structural analysis [@Zienkiewicz1972] with several commercial software being developed since the end of the 20th century. In the last decades new paradigms, such as Building Information Modeling or Parametric Design, have been applied to design in Engineering with increasing attention to openness and automatization. Open-source software (OSS) for structural analysis might represent a relevant asset for researchers and engineers providing solutions in design.
+The design of structures relies on the computation (or estimation) of the stress and deformation developed by structural elements subjected to external loads. The Finite Element Method (FEM) can be considered the most effective computational tool for structural analysis [@Zienkiewicz1972], with several commercial software packages being developed since the end of the 20th century. In recent decades, new paradigms such as Building Information Modeling or Parametric Design have been applied to engineering design, with increasing attention to openness and automation. Open-source software (OSS) for structural analysis might represent a relevant asset for researchers and engineers providing solutions in design.
 
-The main goal of the present library is to provide an open implementation of the FEM for nonlinear structural analysis problems. The library allows any non-expert user to solve static or dynamic problems considering highly nonlinear phenomena such as large rotations and/or nonlinear dynamics. The mathematical formulations implemented are based on relevant references, from classical textbooks [@Bathe1982;@Crisfield1997] to recent journal articles [@Battini2002;@Le2014]. Finally, the library can be executed in any platform supporting GNU-Octave/MATLAB, allowing its integration it with other software in the analysis/design process.
+The main goal of the present library is to provide an open implementation of the FEM for nonlinear structural analysis problems. The library allows any non-expert user to solve static or dynamic problems considering highly nonlinear phenomena such as large rotations and/or nonlinear dynamics. The mathematical formulations implemented are based on relevant references, from classical textbooks [@Bathe1982;@Crisfield1997] to recent journal articles [@Battini2002;@Le2014]. Finally, the library can be executed on any platform supporting GNU-Octave/MATLAB, allowing its integration with other software in the analysis/design process.
 
 # Statement of need
 
-As mentioned above, new innovative and open standards, such as BIM [@IFC] or SAF [@SAF], have emerged in the Engineering Design industries, allowing OSS to be more easily integrated in workflows. Most of OSS FEM software are aimed to efficiently solve general continuum-domain problems without natively considering structural elements [@freefem;@fenics]. For structural analysis, libraries such as [@FEAP;@opensees] or more recently [@cubens], can be found. However their code is written in Fortran, C or C++, which represents a limitation for graduate students or researchers whose expertise and user experience is focused on Octave/Matlab. A structural analysis library in Octave/Matlab could provide a considerable group of Engineering professionals and academics, software tools that are: easy-to-use, extendable and free (to use and inspect).
[…]
