+++ docs/src/references.md
@@ -1,6 +1,13 @@
-### References
+### Theory and References
+
+#### Element Formulations
+
+ONSAS provides a set of finite elements for structural analysis. For 1D structures, truss and frame elements are available, supporting both linear and large-displacement analysis via a co-rotational formulation. The nonlinear static frame formulation is based on (Battini and Pacoste, 2002), while the dynamic implementation follows the consistent co-rotational approach by Le et al. (2014). For continuum problems, 2D (triangle) and 3D (tetrahedron) elements are implemented, allowing for Linear Elastic, Saint-Venant-Kirchhoff, and Neo-Hookean material models. For plane triangle elements plastic analysis in available based on 
+
+#### Key References
 
  * (Bathe, 2014) Klaus-Jurgen Bathe.  Finite Element Procedures . 2014.
  * (Bazzano and Pérez Zerpa, 2017) J. B. Bazzano and J. Perez Zerpa.  Introducción al Análisis No Lineal de Estructuras. 2017.
  * (Battini and Pacoste, 2002) Co-rotational beam elements with warping effects in instability problems, Computer Methods in Applied Mechanics and Engineering, 191 (17-18). 2020.
  * (Holzapfel, 2000) Nonlinear Solid Mechanics, A continuum approach for Engineering, 2000, Wiley.
+ * (Le et.al., 2014) Thanh-Nam Le and Jean-Marc Battini and Mohammed Hjiaj. A consistent 3D corotational beam element for nonlinear dynamic analysis of flexible structures, Computer Methods in Applied Mechanics and Engineering, 269, 2014.
