+++ paper/paper.bib
@@ -281,3 +281,11 @@ keywords = {Reachability analysis, Finite element method, Heat transfer, Structu
   doi          = {10.5281/zenodo.7658871},
   url          = {https://doi.org/10.5281/zenodo.7658871}
 }
+
+@book{deSouzaNeto2008,
+  title={Computational Methods for Plasticity: Theory and Applications},
+  author={de Souza Neto, Eduardo A. and Peri\'{c}, Djordje and Owen, David R. J.},
+  year={2008},
+  publisher={Wiley},
+  isbn={978-0-470-69452-7}
+}
