### 2026-03-02 #9604 @jorgepz
Dear @AnjaliSandip,

I have made several modifications to both the manuscript and the repository to address all the comments. Thank you again to @akchurinda and @odakan for your valuable feedback, which has led to a significantly improved version of both the paper and the library.

The updates to the repository are reflected in the latest release and are detailed in the corresponding [Full Changelog](https://github.com/ONSAS/ONSAS/compare/v0.4.1...v0.4.2). The revisions made to the paper can be tracked in [this PR](https://github.com/ONSAS/ONSAS/pull/794/changes).

Key changes include:

   - Added contribution guidelines: A CONTRIBUTING.md file was added to the repository.

   - Resolved editorial comments: Addressed all editorial feedback, which included rephrasing several citations to better integrate the authors into the flow of the text.

   - Expanded software discussion: Broadened the discussion on structural FEM libraries to include an additional library, and discussing features and the focus of the most relevant libraries.

   - Enhanced documentation and citations: Improved the Reference section in the ONSAS documentation by adding a detailed list of the formulations implemented for each element type, along with their specific references. To support this, key texts such as de Souza Neto's book on computational plasticity were added to both the manuscript and the documentation.

Thanks you,
Jorge
