### 2026-03-09 #9604 @akchurinda
@jorgepz,

Thank you for updating the manuscript. I think the paper is ready for publication after two minor comments are addressed:

1. Fix citation on line 88.
2. I think capitalizations such as "Engineering Design" (line 30) and "structural Engineering" (line 52) are unnecessary, but up to you.

Otherwise, the documentation, `CONTRIBUTING.md`, and other aux. items LGTM.
