### 2026-03-09 72f11f3
Correct minor grammatical errors in paper.md
