+++ paper/paper.md
@@ -57,11 +57,11 @@ The main goal of the present library is to provide an open implementation of the
 
 # Statement of need
 
-As mentioned above, new innovative and open standards, such as BIM [@IFC] or SAF [@SAF], have emerged in the Engineering Design industries, allowing OSS to be more easily integrated in workflows. Most OSS FEM software is aimed at efficiently solving general continuum-domain problems without natively supporting structural elements [@freefem;@fenics]. For structural analysis, libraries such as FEAP [@FEAP], OpenSees [@opensees] or CU-BENs [@cubens], can be found. FEAP is a general-purpose finite element analysis program designed for research and educational use, allowing the solution of problems with linear, surface, and volume elements, including contact and mesh generation, among other features. OpenSees is a software framework for developing applications to simulate the performance of structural and geotechnical systems subjected to earthquakes. Community developments have added features such as fluid-structure interaction and a Python interface.
+As mentioned above, new innovative and open standards, such as BIM [@IFC] or SAF [@SAF], have emerged in the engineering design industries, allowing OSS to be more easily integrated in workflows. Most OSS FEM software is aimed at efficiently solving general continuum-domain problems without natively supporting structural elements [@freefem;@fenics]. For structural analysis, libraries such as FEAP [@FEAP], OpenSees [@opensees] or CU-BENs [@cubens], can be found. FEAP is a general-purpose finite element analysis program designed for research and educational use, allowing the solution of problems with linear, surface, and volume elements, including contact and mesh generation, among other features. OpenSees is a software framework for developing applications to simulate the performance of structural and geotechnical systems subjected to earthquakes. Community developments have added features such as fluid-structure interaction and a Python interface.
 
 The core code of the tools mentioned above is written in Fortran, C, or C++, which represents a limitation for graduate students or researchers whose experience is focused on scripting languages. Structural analysis libraries developed using Octave/Matlab or Python could represent a more suitable choice for Engineering professionals, students, or academics who need to solve problems with certain complex structural behaviors as well as being able to inspect the code and the formulations implemented.
 
-Among the available tools, PyElastica [@PyElastica] can be highlighted, since it is a Python library for simulating assemblies of slender, one-dimensional structures using Cosserat Rod theory. It can model large deformations, include hinges, and also solve problems including contact between slender elements. The tool is focused on one-dimensional elements and does not allow the solution of solid, plane, or shell elements, which are frequent in structural Engineering applications.
+Among the available tools, PyElastica [@PyElastica] can be highlighted, since it is a Python library for simulating assemblies of slender, one-dimensional structures using Cosserat Rod theory. It can model large deformations, include hinges, and also solve problems including contact between slender elements. The tool is focused on one-dimensional elements and does not allow the solution of solid, plane, or shell elements, which are frequent in structural engineering applications.
 
 The development of the library presented in this article, called Open Nonlinear Structural Analysis Solver (`ONSAS`), started in 2017. In this course, graduate students of a Nonlinear Structural Analysis course taught at the School of Engineering of _Universidad de la República_ in Uruguay are introduced to the mathematical formulation of the Principle of Virtual Work (PVW) and its numerical resolution methods for truss structures. The set of functions and scripts developed was published in the course book by @anle.
 
@@ -81,7 +81,7 @@ $$
 $$
[…]
