### 2026-03-10 #9604 @odakan
Dear @jorgepz,

Congratulations on the work; the paper and the ONSAS package are excellent contributions. During my review of the source code, I identified two points that you might want to address to ensure runtime robustness:

1) Function Name Mismatch in assembler.m:
I noticed a potential "undefined function" error for plastic 2D frame elements in dynamic analysis. In assembler.m (Lines 237-240), the code calls:
```matlab
      if dynamicProblemBool
        [fs, ks] = frame_inertial_force(elemNodesxyzRefCoords, elemCrossSecParams, ...
                                          [1 modelParams], elemDisps, ...
                                          dotdispsElem, dotdotdispsElem, ...
                                          density, massMatType, analysisSettings);
```
But the actual function in the frameInertialForce.m file is named `frameInertialForce`. This could trigger an "undefined function" error at runtime.

2) Convergence Logic in convergenceTest.m
In convergenceTest.m (lines 42-44), the logic for reaching the maximum iteration limit is currently:
```matlab
elseif dispIter >= stopTolIts
  stopCritPar = 3;
  booleanConverged = 1;
```
If stopTolIts is the maximum number of iterations allowed, when `dispIter >= stopTolIts` is reached, `booleanConverged` is set to 1. This means the solver exits the iteration loop. If this is the intended behavior to allow the simulation to proceed, I suggest adding a warning message to alert the user that the maximum number of iterations was reached. However, if this is not the intended behavior, you might consider adding step-reduction logic or a hard error when stopCritPar == 3 to prevent the solver from silently accepting unconverged solutions.

I hope these suggestions are helpful for the final polish of the package!

Best,
Deniz
