+++ src/core/assembler.m
@@ -234,7 +234,7 @@ function [fsCell, stressMat, tangMatsCell, localInternalForces, strain_vec, acum
         frameStateCellnp1(elem, :) = aux;
 
         if dynamicProblemBool
-          [fs, ks] = frame_inertial_force(elemNodesxyzRefCoords, elemCrossSecParams, ...
+          [fs, ks] = frameInertialForce(elemNodesxyzRefCoords, elemCrossSecParams, ...
                                           [1 modelParams], elemDisps, ...
                                           dotdispsElem, dotdotdispsElem, ...
                                           density, massMatType, analysisSettings);
