### 2026-03-10 3381d13
Add warning for maximum iterations reached
