### 2026-03-10 aae3aaa
Fix formatting in frameInertialForce function call
