+++ src/core/assembler.m
@@ -235,9 +235,9 @@ function [fsCell, stressMat, tangMatsCell, localInternalForces, strain_vec, acum
 
         if dynamicProblemBool
           [fs, ks] = frameInertialForce(elemNodesxyzRefCoords, elemCrossSecParams, ...
-                                          [1 modelParams], elemDisps, ...
-                                          dotdispsElem, dotdotdispsElem, ...
-                                          density, massMatType, analysisSettings);
+                                        [1 modelParams], elemDisps, ...
+                                        dotdispsElem, dotdotdispsElem, ...
+                                        density, massMatType, analysisSettings);
 
           Fmase = fs{3};
           Ce = ks{2};
