### 2026-03-10 1a578f9
Update convergenceTest.m
