### 2026-03-10 #9604 @jorgepz
Thank you @odakan for addressing those issues!
I've addressed both in [this PR](https://github.com/ONSAS/ONSAS/pull/796).
Best,
Jorge
