### 2026-03-11 #9604 @AnjaliSandip
> [@jorgepz](https://github.com/jorgepz), thank you for a quick update of the paper. This package is a great addition to the tools available to the structural engineers.
>
> [@AnjaliSandip](https://github.com/AnjaliSandip), I recommend that paper to be accepted without any further revisions on my part.

Thank you.
