### 2026-03-12 0b2499f
Modify citation and bump version to 0.4.4
Updated author details and version number, so that it matches authors of the JOSS article.
