+++ CITATION.cff
@@ -7,25 +7,19 @@ authors:
 - family-names: "Vanzulli Pena"
   given-names: "M. C."
   orcid: "https://orcid.org/0000-0001-6080-3815"
-- family-names: "Viera Sosa"
-  given-names: "J."
-  orcid: "https://orcid.org/0000-0003-3615-6966"
-- family-names: "Schaedler de Almeida"
-  given-names: "F."
-  orcid: "https://orcid.org/0000-0003-1535-5405"
-- family-names: "Merlino Chiozza"
-  given-names: "S. A."
 - family-names: "Villié"
   given-names: "A."
   orcid: "https://orcid.org/0000-0001-5431-2892" 
-- family-names: "Forets"
-  given-names: "M."
-  orcid: "https://orcid.org/0000-0002-9831-7801"  
-- family-names: "Battini"
-  given-names: "J.-M."
-  orcid: "https://orcid.org/0000-0003-2104-382X"
+- family-names: "Viera Sosa"
+  given-names: "J."
+  orcid: "https://orcid.org/0000-0003-3615-6966"
+- family-names: "Merlino Chiozza"
+  given-names: "S. A."
+- family-names: "Schaedler de Almeida"
+  given-names: "F."
+  orcid: "https://orcid.org/0000-0003-1535-5405"
 - family-names: "Bazzano García"
   given-names: "J. B."
 title: "ONSAS: an Open Nonlinear Structural Analysis Solver for GNU-Octave/Matlab"
-version: 0.4.1
+version: 0.4.4
 url: "https://github.com/ONSAS/ONSAS"
