### 2026-03-12 #9604 @jorgepz
Thanks for your message @AnjaliSandip.

The current version is: 0.4.4 and the zenodo DOI is: 10.5281/zenodo.18973164

Regarding the missing DOI of Bathe's book it seems that editorial bot is suggesting adding [the DOI of a REVIEW of the book](https://www.sciencedirect.com/science/article/abs/pii/0141029684900142?via%3Dihub) that was published in a journal, but not the Finite Element Procedures book. I did not find a DOI of the cited 1982 Bathe's book.
