### 2026-03-12 fc193a2
Update contributor list in index.md
Added J. Bruno Bazzano as a relevant contributor for buckling analysis modules and validation examples.
