### 2026-03-12 51e36d6
Update contributor list and add new affiliations
