### 2026-03-12 05970c3
Update contributors and add new references
