### 2026-03-16 335ef31
Remove Battini and Pacoste article from bibliography
Removed outdated article reference by Battini and Pacoste.
