+++ paper/paper.bib
@@ -21,18 +21,6 @@ title = {Introductory Lectures on the Finite Element Method},
 year = {1972}
 }
 
-@article{Battini2002,
-author = {Battini, Jean Marc and Pacoste, Costin},
-doi = {10.1016/S0045-7825(01)00352-8},
-issn = {00457825},
-journal = {Computer Methods in Applied Mechanics and Engineering},
-number = {17-18},
-pages = {1755--1789},
-title = {Co-rotational beam elements with warping effects in instability problems},
-volume = {191},
-year = {2002}
-}
-
 @article{yoshi,
 title = {Elastic buckling phenomenon applicable to deployable rings},
 journal = {International Journal of Solids and Structures},
