### 2026-03-16 #9604 @jorgepz
Dear @AnjaliSandip, thanks for your work. I removed one of the two bib entries for the article Battini2002. Now it should be working.
