### 2026-03-16 #9604 @jorgepz
This might not be very important, but I suggest considering removing the _M_ language label for this issue, for the same reasons I mentioned in https://github.com/openjournals/joss-reviews/issues/8473#issuecomment-3136924915.
