### 2026-03-17 #9604 @jorgepz
> Dear [@AnjaliSandip](https://github.com/AnjaliSandip), thanks for your work. I removed one of the two bib entries for the article Battini2002. Now it should be working.
>

This was done in [this commit](https://github.com/ONSAS/ONSAS/commit/335ef31d742662799840025748fb6a5e3c9a2611)
