### 2026-03-19 032f014
Clarify PyElastica's capabilities and limitations
Reworded the description of PyElastica to clarify its focus and limitations in structural engineering applications.
