+++ paper/paper.md
@@ -61,7 +61,7 @@ As mentioned above, new innovative and open standards, such as BIM [@IFC] or SAF
 
 The core code of the tools mentioned above is written in Fortran, C, or C++, which represents a limitation for graduate students or researchers whose experience is focused on scripting languages. Structural analysis libraries developed using Octave/Matlab or Python could represent a more suitable choice for Engineering professionals, students, or academics who need to solve problems with certain complex structural behaviors as well as being able to inspect the code and the formulations implemented.
 
-Among the available tools, PyElastica [@PyElastica] can be highlighted, since it is a Python library for simulating assemblies of slender, one-dimensional structures using Cosserat Rod theory. It can model large deformations, include hinges, and also solve problems including contact between slender elements. The tool is focused on one-dimensional elements and does not allow the solution of solid, plane, or shell elements, which are frequent in structural engineering applications.
+Among the available tools, PyElastica [@PyElastica] can be highlighted, since it is a Python library for simulating assemblies of slender, one-dimensional structures using Cosserat Rod theory. It can model large deformations, include hinges, and also solve problems including contact between slender elements. The tool focuses on one-dimensional elements and does not support the analysis of solid, plane, or shell elements, which are common in structural engineering applications.
 
 The development of the library presented in this article, called Open Nonlinear Structural Analysis Solver (`ONSAS`), started in 2017. In this course, graduate students of a Nonlinear Structural Analysis course taught at the School of Engineering of _Universidad de la República_ in Uruguay are introduced to the mathematical formulation of the Principle of Virtual Work (PVW) and its numerical resolution methods for truss structures. The set of functions and scripts developed was published in the course book by @anle.
 
