+++ paper/paper.md
@@ -63,7 +63,7 @@ The core code of the tools mentioned above is written in Fortran, C, or C++, whi
 
 Among the available tools, PyElastica [@PyElastica] can be highlighted, since it is a Python library for simulating assemblies of slender, one-dimensional structures using Cosserat Rod theory. It can model large deformations, include hinges, and also solve problems including contact between slender elements. The tool focuses on one-dimensional elements and does not support the analysis of solid, plane, or shell elements, which are common in structural engineering applications.
 
-The development of the library presented in this article, called Open Nonlinear Structural Analysis Solver (`ONSAS`), started in 2017. In this course, graduate students of a Nonlinear Structural Analysis course taught at the School of Engineering of _Universidad de la República_ in Uruguay are introduced to the mathematical formulation of the Principle of Virtual Work (PVW) and its numerical resolution methods for truss structures. The set of functions and scripts developed was published in the course book by @anle.
+The development of the library presented in this article, called Open Nonlinear Structural Analysis Solver (`ONSAS`), started in 2017. In this course, graduate students of a Nonlinear Structural Analysis course taught at the School of Engineering of _Universidad de la República_ in Uruguay are introduced to the mathematical formulation of the Principle of Virtual Work (PVW) and its numerical resolution methods for truss structures. The initial subset of functions and scripts was published in the course book by @anle.
 
 Subsequent development of the code was mostly motivated by the research projects that users worked on, and as the complexity of the problems posed increased, nonlinear analysis of frame structural elements became necessary. In 2019, an implementation of the co-rotational formulation for nonlinear static analysis  presented by @Battini2002, was contributed by Prof. Battini. During 2022, the consistent co-rotational formulation for dynamic analysis presented by @Le2014 was implemented. To the best knowledge of the authors, this library represents the first open implementation of the cited dynamic co-rotational formulation. By the end of 2022, the nonlinear dynamic co-rotational formulation was applied to include aerodynamic loads in frames, allowing the solution of a new set of modeling problems such as wind turbine aerodynamic analysis, as shown by @mauri. More recently, a Vortex-Induced Vibrations model for fluid-structure interaction was included and applied in @villie. Currently, in the context of a Master's theses, a nonlinear shell element formulation is being implemented.
 
