### 2026-03-19 f95411e
Refine paper content by removing redundancy
Removed redundant sentence about ONSAS applications.
