+++ dev/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"references/#Theory-and-References","page":"References","title":"Theory and References","text":"","category":"section"},{"location":"references/#Element-Formulations","page":"References","title":"Element Formulations","text":"ONSAS provides a diverse set of finite element formulations for structural analysis. We strongly encourage users to explore the references used to implement each formulation.\n\nFor 1D structures, truss and frame elements are available, supporting both linear and large-displacement analysis via a co-rotational formulation. The nonlinear static truss co-rotational implementation is based on (Crisfield, 1997). The nonlinear static frame formulation is based on (Battini and Pacoste, 2002), while the dynamic implementation follows the consistent co-rotational approach proposed by Le et al. (2014). For planar frame problems, softening hinges are also implemented based on (Jukic, 2013).\n\nFor continuum problems, 2D (triangle) and 3D (tetrahedron) elements are implemented, supporting Linear Elastic, Saint-Venant-Kirchhoff, and Neo-Hookean material models. For plane triangle elements, elasto-plastic analysis is available, utilizing a formulation based on (de Souza Neto et al., 2008). ","category":"section"},{"location":"references/#Key-References","page":"References","title":"Key References","text":"(Bathe, 2014) Klaus-Jurgen Bathe.  Finite Element Procedures . 2014.\n(Bazzano and Pérez Zerpa, 2017) J. B. Bazzano and J. Perez Zerpa.  Introducción al Análisis No Lineal de Estructuras. 2017.\n(Battini and Pacoste, 2002) Co-rotational beam elements with warping effects in instability problems, Computer Methods in Applied Mechanics and Engineering, 191 (17-18). 2002.\n(Crisfield, 1997) M. A. Crisfield, Non-Linear Finite Element Analysis Solids and Structure, Volume 2, Advanced Topics, , Wiley, 1997.\n(Holzapfel, 2000) G. Holzapfel, Nonlinear Solid Mechanics, A continuum approach for Engineering, 2000, Wiley.\n(Jukic, 2013) M. Jukić and B. Brank and A. Ibrahimbegović, Embedded discontinuity finite element formulation for failure analysis of planar reinforced concrete beams and frames, Engineering Structures, 50, 2013.\n(Le et.al., 2014) Thanh-Nam Le and Jean-Marc Battini and Mohammed Hjiaj. A consistent 3D corotational beam element for nonlinear dynamic analysis of flexible structures, Computer Methods in Applied Mechanics and Engineering, 269, 2014.\n(de Souza Neto, et.al., 2008) E. A. de Souza Neto and D. Perić and D. R. J. Owen, Computational Methods for Plasticity: Theory and Applications, 2008, Wiley.","category":"section"},{"location":"examples/springMass/#Spring-mass-damper-system-example","page":"Spring-mass system","title":"Spring-mass-damper system example","text":"(Image: Octave script)","category":"section"},{"location":"examples/springMass/#ODE-System-definition","page":"Spring-mass system","title":"ODE System definition","text":"In this example a simple spring-mass-damper system is considered. This simple problem is used to validate different implementations of truss and frame elements submitted to dynamic loads.\n\n<img src=\"../../assets/springMassSystem.svg\" alt=\"spring-mass diagram\" width=\"800\"/>","category":"section"},{"locat
[…]
