+++ versions.js
@@ -1,0 +1,9 @@
+var DOC_VERSIONS = [
+  "stable",
+  "v0.4",
+  "v0.3",
+  "v0.2",
+  "dev",
+];
+var DOCUMENTER_NEWEST = "v0.4.4";
+var DOCUMENTER_STABLE = "stable";

[…]
