### 2026-03-23 #9604 @AnjaliSandip
@kyleniemeyer, do you have any suggestions on how we could resolve this issue?

> Dear [@AnjaliSandip](https://github.com/AnjaliSandip) in [this PR](https://github.com/ONSAS/ONSAS/pull/797/changes) I did a few minor changes in the text of the paper. I did not make any changes to the code.
>
Regarding the "MISSING DOIs", I was able to verify that the DOI 10.1016/0141-0296(84)90014-2 that the editorial bot suggests for Bathe's book is a REVIEW of the book, not the book itself. The book does not have a DOI. I am not aware of how to tell the editorial bot to avoid suggesting a DOI for that book
