### 2026-03-28 #9604 @jorgepz
@AnjaliSandip, while we wait for assistance, I guess that the PR https://github.com/openjournals/joss-papers/pull/7927 was closed due to my last change in the proof paper, and the missing doi bot warning did not caused that.

Would the process be restarted if you recommend-accept again?
