### 2026-03-30 #9604 @AnjaliSandip
> [@AnjaliSandip](https://github.com/AnjaliSandip), while we wait for assistance, I guess that the PR [openjournals/joss-papers#7927](https://github.com/openjournals/joss-papers/pull/7927) was closed due to my last change in the proof paper, and the missing doi bot warning did not caused that.
>
> Would the process be restarted if you recommend-accept again?

Yes, is there another related reference that you could use instead?
