### 2026-03-31 ffc3732
Revise references in FEM library summary
Updated references in the summary section to include relevant sources for FEM implementation.
