+++ paper/paper.md
@@ -53,7 +53,7 @@ bibliography: paper.bib
 
 The design of structures relies on the computation (or estimation) of the stress and deformation developed by structural elements subjected to external loads. The Finite Element Method (FEM) can be considered the most effective computational tool for structural analysis [@Zienkiewicz1972], with several commercial software packages being developed since the end of the 20th century. In recent decades, new paradigms such as Building Information Modeling or Parametric Design have been applied to engineering design, with increasing attention to openness and automation. Open-source software (OSS) for structural analysis might represent a relevant asset for researchers and engineers providing solutions in design.
 
-The main goal of the present library is to provide an open implementation of the FEM for nonlinear structural analysis problems. The library allows any non-expert user to solve static or dynamic problems considering highly nonlinear phenomena such as large rotations and/or nonlinear dynamics. The mathematical formulations implemented are based on relevant references, from classical textbooks [@Bathe1982;@Crisfield1997] to recent journal articles [@Battini2002;@Le2014]. Finally, the library can be executed on any platform supporting GNU-Octave/MATLAB, allowing its integration with other software in the analysis/design process.
+The main goal of the present library is to provide an open implementation of the FEM for nonlinear structural analysis problems. The library allows any non-expert user to solve static or dynamic problems considering highly nonlinear phenomena such as large rotations and/or nonlinear dynamics. The mathematical formulations implemented are based on relevant references, from classical textbooks [@Crisfield1997] to recent journal articles [@Battini2002;@Le2014]. Finally, the library can be executed on any platform supporting GNU-Octave/MATLAB, allowing its integration with other software in the analysis/design process.
 
 # Statement of need
 
