### 2026-03-31 68af9a5
Delete Bathe1982 entry from paper.bib
Removed Bathe1982 entry from bibliography.
