### 2026-03-31 #9604 @jorgepz
Hi, in https://github.com/ONSAS/ONSAS/pull/798 I removed the reference. Now I understand there are no issues left. thanks again.
