### 2026-03-31 #9604 @kyleniemeyer
Hi @jorgepz, just a few edits needed before accepting:
- In the "Features and examples" section, please edit the citation command used in "based on (Souza Neto et al., 2008)" from `[@author:2001]` to `@author:2001` to fix the in-text formatting to "based on Souza Neto et al. (2008)".
- The Zienkiewicz 1972 article is missing some bibliographic information
