### 2026-04-01 d48fb8e
Add new book entry for Finite Element Procedures
