+++ paper/paper.bib
@@ -34,6 +34,15 @@ url = {https://www.sciencedirect.com/science/article/pii/002076839290024N},
 author = {Goto Yoshiaki and Watanabe Yasuhito and Kasugai Toshihiro and Obata Makoto}
 }
 
+@book{Bathe1982,
+author = {Bathe, Klaus-J{\"{u}}rgen},
+isbn = {978-0133173055},
+pages = {736},
+publisher = {Prentice-Hall},
+title = {{Finite Element Procedures in Engineering Analysis}},
+year = {1982}
+}
+
 @book{Crisfield1997,
 annote = {0020},
 author = {Crisfield, Mike A.},
