### 2026-04-01 8a5e08b
Update references in FEM library description
