+++ paper/paper.md
@@ -53,7 +53,7 @@ bibliography: paper.bib
 
 The design of structures relies on the computation (or estimation) of the stress and deformation developed by structural elements subjected to external loads. The Finite Element Method (FEM) can be considered the most effective computational tool for structural analysis [@Zienkiewicz1972], with several commercial software packages being developed since the end of the 20th century. In recent decades, new paradigms such as Building Information Modeling or Parametric Design have been applied to engineering design, with increasing attention to openness and automation. Open-source software (OSS) for structural analysis might represent a relevant asset for researchers and engineers providing solutions in design.
 
-The main goal of the present library is to provide an open implementation of the FEM for nonlinear structural analysis problems. The library allows any non-expert user to solve static or dynamic problems considering highly nonlinear phenomena such as large rotations and/or nonlinear dynamics. The mathematical formulations implemented are based on relevant references, from classical textbooks [@Crisfield1997] to recent journal articles [@Battini2002;@Le2014]. Finally, the library can be executed on any platform supporting GNU-Octave/MATLAB, allowing its integration with other software in the analysis/design process.
+The main goal of the present library is to provide an open implementation of the FEM for nonlinear structural analysis problems. The library allows any non-expert user to solve static or dynamic problems considering highly nonlinear phenomena such as large rotations and/or nonlinear dynamics. The mathematical formulations implemented are based on relevant references, from classical textbooks [@Bathe1982;@Crisfield1997] to recent journal articles [@Battini2002;@Le2014]. Finally, the library can be executed on any platform supporting GNU-Octave/MATLAB, allowing its integration with other software in the analysis/design process.
 
 # Statement of need
 
@@ -81,7 +81,7 @@ $$
 $$
 where $\textbf{f}_{int}$ and $\textbf{f}_{vis}$ are the vectors of internal static and damping forces, $\textbf{f}_{ext,t}$ is the vector of external forces, $\textbf{f}_{ext,add}$ is the vector of external additional forces (caused by different external agents) that can be set by the user, and $\textbf{f}_{ine}$ is the vector of inertial forces. The vectors $\textbf{u}_t$, $\dot{\textbf{u}}_t$, and $\ddot{\textbf{u}}_t$ represent the displacements, velocities, and accelerations of all the degrees of freedom of the structure at time $t$, respectively.
 
-For the numerical time integration, `ONSAS` includes built-in nonlinear solution strategies, such as the Newton-Raphson and the Arc-Length methods, for static analyses, and the Newmark and the $\alpha$-HHT methods for transient dynamic analyses. Linear modal analysis is available for frame and truss elements. For solid elements, it is possible to compute the tangent matrix associated with the internal forces by using the complex-step approach presented in @kiran. For planar frame structures it is also possible to perform plastic analyses considering softening hinges as described in @jukic, while for plane triangular elements elasto-plastic analyses based on [@deSouzaNeto2008] are also available.
+For the numerical time integration, `ONSAS` includes built-in nonlinear solution strategies, such as the Newton-Raphson and the Arc-Length methods, for static analyses, and the Newmark and the $\alpha$-HHT methods for transient dynamic analyses. Linear modal analysis is available for frame and truss elements. For solid elements, it is possible to compute the tangent matrix associated with the internal forces by using the complex-step approach presented in @kiran. For planar frame structures it is also possible to perform plastic analyses considering softening hinges as described in @jukic, while for plane triangular elements elasto-plastic analyses based on @deSouzaNeto2008 are also available.
 
[…]
