### 2026-04-01 92fa5cf
Add publisher and pages to a bibliography entry
