+++ paper/paper.bib
@@ -18,6 +18,8 @@ author = {Zienkiewicz, Olgierd C.},
 booktitle = {Introductory Lectures on the Finite Element Method},
 doi = {10.1007/978-3-7091-2973-9},
 title = {Introductory Lectures on the Finite Element Method},
+publisher = {Springer Vienna},
+pages = {99},
 year = {1972}
 }
 
