### 2026-04-01 #9604 @jorgepz
thank you @kyleniemeyer . Considering your comments, I did three changes in https://github.com/ONSAS/ONSAS/pull/799/changes:
 - I changed the de souza neto citation.
 - I added two more fields to the Zienkiewicz book.
 - based on your comment about the false positive, I added back again the Bathe's book citation since it is a fundamental reference in the area.
