### 2026-04-01 #9604 @kyleniemeyer
@jorgepz Thanks - though I noticed one other issue: the "Group, N (2023)" citation, from `.bib` entry `SAF`. Please wrap the organization name with extra `{}` so that it isn't interpreted as a person's name.
