### 2026-04-01 8e1f1c7
Fix author formatting in SAF entry
