+++ paper/paper.bib
@@ -191,7 +191,7 @@ journal = {Journal of Open Source Software}
 
 @misc{SAF,
 	title = {Structural Analysis Format documentation},
-  author= {Nemetschek~Group},
+  author= {{Nemetschek~Group}},
   year={2023},
 	url = {https://www.saf.guide/en/stable/}
 }
