### 2026-04-04 #9604 @jorgepz
> Thank you [@akchurinda](https://github.com/akchurinda), in [this commit](https://github.com/ONSAS/ONSAS/commit/72f11f3caf7e69c81f042c1540ca059d7e012217) I've addressed the comments. Thanks again for your work.
>
> Dear [@AnjaliSandip](https://github.com/AnjaliSandip), while we wait for [@odakan](https://github.com/odakan) 's re-confirmation, I suggest considering removing the M language label for this issue, for the same reasons I mentioned in [the PRE-REVIEW](https://github.com/openjournals/joss-reviews/issues/8473#issuecomment-3136924915).
>
> Best, Jorge

@kyleniemeyer @AnjaliSandip  I can see that in the final publication of ONSAS https://joss.theoj.org/papers/10.21105/joss.09604 the "M" label is applied. I mention this just in case it is possible to manually remove it since M code is not used in ONSAS and this is due to a github error. Thanks
