### 2026-04-05 #9604 @kyleniemeyer
Hm, it looks like the `M` tag is still applied. @openjournals/dev @arfon any idea how to fix this?
